% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataREACH.R
\name{dataREACH}
\alias{dataREACH}
\title{dataREACH: Process Project REACH data}
\usage{
dataREACH(
  base_wd,
  overwrite = FALSE,
  proc_source = FALSE,
  data_list = "all",
  micro_protocols = "all",
  micro_data_type = "all",
  overwrite_ggir_derivs = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{proc_source}{whether to processes the raw data using proc_tasks.R. Default = FALSE}

\item{data_list}{list of strings matching the notes below to indicate the data to be written. Default = 'all' to export all data and metadata. Options include:
\itemize{
\item{'participants' - BIDS specified participants.tsv file}
\item{'actigraph' - activity and sleep data generated from GGIR and mMARCH.AC}
\item{'microstructure' - coded meal microstructure data}
\item{'foodview' - fMRI Food Viewing task}
\item{'nihtoolbox' - NIH Toolbox data}
\item{'pit' - Pavlovian Instrumental Transfer task data}
\item{'rrv' - Relative Reinforcing Value of Food task}
\item{'spacegame' - Space Game data (need to finish processing in Matlab)}
\item{'sst' - fMRI Stop-Signal Task data}
\item{'anthropometrics' - height, weight, and computed anthropometric data}
\item{'demographics' - compiled demographic data}
\item{'dxa' - verified DXA data}
\item{'household' - compiled demographicinformation about houshold}
\item{'infancy' - compiled demographic information related to infancy}
\item{'intake' - compiled intake data with computed intake values}
\item{'mri_visit' - MRI visit information including Freddy and CAMS}
\item{'parent_updates' - all visit updates}
\item{'researcher_notes' - all visit notes}
\item{'audit' - Alcohol Use Disorders Identification Test}
\item{'bes' - Binge Eating Scale}
\item{'bisbas' - Behavioral Inhibition System/Behavioral Activation System}
\item{'brief2' - Behavioral Rating Inventory of Executive Function-2}
\item{'cbq' - Child Behavior Questionnaire}
\item{'cchip' - Community Childhood Hunger Identification Project}
\item{'cebq' - Children's Eating Behavior Questionnaire}
\item{'cfpq' - Comprehensive Feeding Practices Questionnaire}
\item{'cfq' - Child Feeding Questionnaire}
\item{'chaos' - Confusion, Hubbub, and Order Scale}
\item{'class' - \emph{need}}
\item{'cshq' - Children Sleep Habits Questionnaire}
\item{'debq' - Dutch Eating Behavior Questionnaire}
\item{'efcr' - External Food Cue Responsiveness Scale}
\item{'ffbs' - Family Food Behavior Survey}
\item{'fsq' - \emph{need}}
\item{'hfi' - Fulkerson Home Food Inventory}
\item{'hfias' - Household Food Insecurity Access Scale}
\item{'hfssm' - U.S. Household Food Security Survey Module}
\item{'kbas' - Kid's Brand Awareness Scale}
\item{'lbc' - Lifestyle Behavior Checklist}
\item{'loc' - Loss of Control-Eating Questionnaire}
\item{'pmum' - Problematic Media Use Measure \emph{need}}
\item{'pptq' - Pictorial Personality Traits Questionnaire for Children}
\item{'pss' - Perceived Stress Scale}
\item{'pstca' - \emph{need}}
\item{'puberty' - combination of Tanner and Pubertal Rating Scale}
\item{'pwlb' - Parent Weight-Loss Behavior Questionnaire}
\item{'rank' - Parent ranking of foods sources? \emph{need}}
\item{'scpf' - tructure and Control in Parent Feeding Questionnaire}
\item{'sic' - Stress in Children Questionnaire \emph{need}}
\item{'sleeplog' - Week long sleep log}
\item{'spsrq' - Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
\item{'stq' - Screen Time Questionnaire \emph{need}}
\item{'tfeq' - Three Factor Eating Questionnaire}
}}

\item{micro_protocols}{(optional) list of strings matching the notes below to indicate the which microstructure data. Default = 'all' to export all data and metadata. Options include:
\itemize{
\item{'meal-boring' - meal microstructure behavior after boring commercials}
\item{'meal-food' - meal microstructure behavior after food commercials}
\item{'meal-toy' - meal microstructure behavior after toy commercials}
\item{'eah-boring' - EAH microstructure behavior after boring commercials}
\item{'eah-food' - EAH microstructure behavior after food commercials}
\item{'eah-toy' - EAH microstructure behavior after toy commercials}
}}

\item{micro_data_type}{(optional) Type of data to process for meal microstructure - list of strings matching the data types listed below. Default = 'all' to export both:
\itemize{
\item{'beh_wide' - summary behavioral measures in wide formate by coder. Note: this will write out a summary dataset in bids/phenotype.}
\item{'events_long' - event level data in log format by coder. Note: this writes out a file per participant into bids/rawdata.}
}}

\item{overwrite_ggir_derivs}{(required if 'actigraph' is in data_list or data_list = 'all') overwrite GGIR derivatives. Defult = FLASE, this will take a LONG time.}
}
\value{
Does not return anything
}
\description{
This function:
\itemize{Calls core proc_\* and write_\* functions and writes BIDS-compliant .tsv and .json files into bids/phenotype:
\item{write_redcap: processes and writes out specified REDCap data}
\item{proc_microstructure: processes and writes out microstructure data}
\item{proc_tasks:}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
dataREACH(base_wd, overwrite = FALSE, data_list = 'all')

}

}
