% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_rrv.R
\name{deriv_rrv}
\alias{deriv_rrv}
\title{deriv_rrv: Generate derivative databases for RRV analyses}
\usage{
deriv_rrv(data, file_list)
}
\arguments{
\item{data}{a list of dataframes, where 1 dataframe contains processed RRV data for 1 sub (output of util_task_rrv). data OR file_list is required.}

\item{file_list}{a list filenames, where 1 filename is full path to processed RRV data for 1 sub (exported by util_task_rrv). data OR file_list is required.}
}
\value{
a list with: 1) a dataframe with long summary data (by reinforcer schedule) and 2) a dataframe with overall summary data
}
\description{
This function generates RRV derivative databases from participant-level RRV data
}
\examples{

\dontrun{

# process task data - this returns and exports processed RRV data
base_dir = "/Users/baf44/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/"
task_data <- proc_task(base_wd = base_dir, return_data = TRUE)

# create deriv database from data RETURNED by proc_task
rrv_summary <- deriv_rrv(data = task_data$rrv)

# create deriv database from files EXPORTED by proc_task, using a list of file names
file_list <- list.files(file.path(base_dir, "bids", "rawdata"), pattern = "task-rrv_beh.tsv", recursive = TRUE, full.names = TRUE)
rrv_summary <- deriv_rrv(file_list = file_list)

}
}
