% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_spacegame.R
\name{deriv_spacegame}
\alias{deriv_spacegame}
\title{deriv_spacegame: Get summary data from the Space Game (2-stage reinforcement learning task)}
\usage{
deriv_spacegame(data_list, ses, base_wd, overwrite = FALSE)
}
\arguments{
\item{data_list}{A data frame with variable 'sub_str' that includes all participants}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function calculates summary performance data for a participant and saves the output in a wide format (overall task) and long format (by block)
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.

Function adapted from Alaina Pearce's util_group_spacegame.R script in dataBRAKEr package (https://github.com/alainapearce/dataBRAKEr/blob/main/R/util_group_spacegame.R)
}
\examples{

# process task data for the Food Choice Task
util_group_spacegame <- util_task_spacegame(data_list, ses, data_path, return = TRUE)

\dontrun{
}


}
