% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_sst.R
\name{deriv_sst}
\alias{deriv_sst}
\title{deriv_sst: Generate derivative behavioral databases for SST analyses}
\usage{
deriv_sst(data, file_list)
}
\arguments{
\item{data}{a list of list of dataframes. The top-level list represents individual subjects, and each subject has a sublist of dataframes. Each dataframe contains events data for a specific run of the foodview task for a given sub. A suitable list is returned by proc_task, or can be gathered from files in bids/rawdata}
}
\value{
a list with: 1) summary_long_by_run = a long dataframe with summary data by run and 2) summary_long_by_block = a long dataframe with summary data by block
}
\description{
This function generates SST derivative databases from participant-level beh and func SST files
}
\examples{

\dontrun{

# process task data
base_dir = "/Users/baf44/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/"
task_data <- proc_task(base_wd = base_dir, return_data = TRUE)

# create deriv database from data RETURNED by proc_task
sst_summary <- deriv_sst(data = task_data$sst)

# create deriv database from files EXPORTED by proc_task, using a list of file names
file_list <- list.files(file.path(base_dir, "bids", "rawdata"), pattern = "sst.*\\\\.tsv$", recursive = TRUE, full.names = TRUE) # include files across all subs/modalities/runs
sst_summary <- deriv_sst(file_list = file_list)

}
}
