% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_summary.R
\name{get_participants_summary}
\alias{get_participants_summary}
\title{get_participants_summary:}
\usage{
get_participants_summary(participants_data)
}
\arguments{
\item{participants_data}{(dataframe) particpants_data generated by proc_redcap}
}
\description{
get_participants_summary:
}
