% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_anthropometrics.R
\name{json_anthropometrics}
\alias{json_anthropometrics}
\title{json_anthropometrics: Generates a json file for anthropometric data}
\usage{
json_anthropometrics()
}
\value{
A string with data stored in JSON format containing meta-data anthropometric data
}
\description{
This function generates a json file for anthropometric data
}
