% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_audit.R
\name{json_audit}
\alias{json_audit}
\title{json_audit: Generates a json file for the Alcohol Use Disorders Identification Test}
\usage{
json_audit()
}
\value{
A string with data stored in JSON format containing meta-data the Alcohol Use Disorders Identification Test
}
\description{
This function generates a json file for the scored Alcohol Use Disorders Identification Test and raw participant responses.
}
