% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_bisbas.R
\name{json_bisbas}
\alias{json_bisbas}
\title{json_bisbas: Generates a json file for the Behavioral Inhibition System/Behavioral Activation System Scale}
\usage{
json_bisbas()
}
\value{
A string with data stored in JSON format containing meta-data for the Behavioral Inhibition System/Behavioral Activation System Scale
}
\description{
This function generates a json file for the scored Behavioral Inhibition System/Behavioral Activation System Scale and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_bisbas function in dataprepr and is only accurate for data collected in Study REACH
}
