% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_brief2.R
\name{json_brief2}
\alias{json_brief2}
\title{json_brief2: Generates a json file for the Behavioral Rating Inventory of Executive Function-2}
\usage{
json_brief2()
}
\value{
A string with data stored in JSON format containing meta-data for the Behavioral Rating Inventory of Executive Function-2
}
\description{
This function generates a json file for the scored Behavioral Rating Inventory of Executive Function-2 and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_brief2 function in dataprepr and is only accurate for data collected in Study REACH
}
