% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cbq.R
\name{json_cbq}
\alias{json_cbq}
\title{json_cbq: Generates a json file for the Child Behavior Questionnaire}
\usage{
json_cbq()
}
\value{
A json file documenting the raw inputs and scored values for the Child Behavior Questionnaire
}
\description{
This function generates a json file for the scored Child Behavior Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_cbq function in dataprepr and is only accurate for data collected in Study BREAK
}
