% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cfq.R
\name{json_cfq}
\alias{json_cfq}
\title{json_cfq: Generates a json file for the Child Feeding Questionnaire}
\usage{
json_cfq()
}
\value{
A json file documenting the raw inputs and scored values for the Child Feeding Questionnaire
}
\description{
This function generates a json file for the scored Child Feeding Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_cfq function in dataprepr and is only accurate for data collected in Study REACH
}
