% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_chaos.R
\name{json_chaos}
\alias{json_chaos}
\title{json_chaos: Generates a json file for the Confusion, Hubbub, and Order Scale}
\usage{
json_chaos()
}
\value{
A string with data stored in JSON format containing meta-data for the Confusion, Hubbub, and Order Scale
}
\description{
This function generates a json file for the scored Confusion, Hubbub, and Order Scale and raw participant responses.
}
