% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cshq.R
\name{json_cshq}
\alias{json_cshq}
\title{json_cshq: Generates a json file for the Children\'s Sleep Habits Questionnaire}
\usage{
json_cshq()
}
\value{
A string with data stored in JSON format containing meta-data for the Children\'s Sleep Habits Questionnaire
}
\description{
This function generates a json file for the scored Children\'s Sleep Habits Questionnaire and raw participant responses.
}
