% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_dataset_desc.R
\name{json_dataset_desc}
\alias{json_dataset_desc}
\title{json_dataset_desc: Generate dataset_description JSON for phenotype}
\usage{
json_dataset_desc()
}
\value{
A string with data stored in JSON format containing meta-data for phenotype files derived by dataREACHr
}
\description{
This function generates a json file for phenotype/
}
