% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_debq.R
\name{json_debq}
\alias{json_debq}
\title{json_debq: Generates a json file for the Dutch Eating Behavior Questionnaire}
\usage{
json_debq()
}
\value{
A string with data stored in JSON format containing meta-data for the Dutch Eating Behavior Questionnaire
}
\description{
This function generates a json file for the scored Dutch Eating Behavior Questionnaire and raw participant responses.
}
