% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_demographics.R
\name{json_demographics}
\alias{json_demographics}
\title{json_demographics: Generates a json file for demographics.tsv}
\usage{
json_demographics()
}
\value{
A string with data stored in JSON format containing meta-data for demographic variables
}
\description{
This function generates a json file for demographics.tsv
}
