% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_deriv_beh_dataset_desc.R
\name{json_deriv_beh_dataset_desc}
\alias{json_deriv_beh_dataset_desc}
\title{json_deriv_beh_dataset_desc: Generate dataset_description JSON for beh derivatives}
\usage{
json_deriv_beh_dataset_desc()
}
\value{
A string with data stored in JSON format containing meta-data for beh summary datasets derived by dataREACHr
}
\description{
This function generates a json file for derivatives/dataREACHr
}
