% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_deriv_foodview.R
\name{json_deriv_foodview}
\alias{json_deriv_foodview}
\title{json_deriv_foodview: Generates json file for the derivative Foodview data}
\usage{
json_deriv_foodview()
}
\value{
A list of 2 strings (labels: foodview_bycond_json, foodview_byblock_json) with data stored in JSON format containing meta-data the summary Foodview database
}
\description{
This function generates 2 json files for FoodView summary data reported in derivatives/beh_databases: (1) summary metrics computed by ad condition, (2) summary metrics computed by block
}
