% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_deriv_sst.R
\name{json_deriv_sst}
\alias{json_deriv_sst}
\title{json_deriv_sst: Generates json file for the derivative SST behavioral data}
\usage{
json_deriv_sst()
}
\value{
A list of 2 strings (labels: XX, sst_byblock_json) with data stored in JSON format containing meta-data the summary SST database
}
\description{
This function generates 2 json files for SST summary data reported in derivatives/beh_databases: (1) summary metrics computed by ad condition, (2) summary metrics computed by block
}
