% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_dexa.R
\name{json_dexa}
\alias{json_dexa}
\title{json_dexa: Generates a json file for the dexa}
\usage{
json_dexa()
}
\value{
A string with data stored in JSON format containing meta-data for dexa data
}
\description{
This function generates a json file for XXXX
}
