% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_dxa.R
\name{json_dxa}
\alias{json_dxa}
\title{json_dxa: Generates a json file for the dxa}
\usage{
json_dxa()
}
\value{
A string with data stored in JSON format containing meta-data for dxa data
}
\description{
This function generates a json file for dxa data
}
