% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_eah_wanting.R
\name{json_eah_wanting}
\alias{json_eah_wanting}
\title{json_eah_wanting: Generates a json file for the Eating in the Absence of Hunger (EAH) Wanting Questionnaire}
\usage{
json_eah_wanting()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for the Absence of Hunger (EAH) Wanting Questionnaire
}
