% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_efcr.R
\name{json_efcr}
\alias{json_efcr}
\title{json_efcr: Generates a json file for the External Food Cue Responsiveness Scale}
\usage{
json_efcr()
}
\value{
A json file documenting the raw inputs and scored values for the External Food Cue Responsiveness Scale
}
\description{
This function generates a json file for the scored External Food Cue Responsiveness Scale and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_efcr function in dataprepr and is only accurate for data collected in Study REACH
}
