% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodview.R
\name{json_foodview}
\alias{json_foodview}
\title{json_foodview: Generates a json file for the food view fmri events (data in rawdata/func)}
\usage{
json_foodview()
}
\value{
A string with data stored in JSON format containing meta-data for the food view task
}
\description{
This function generates a json file for cleaned food view task data collected during fmri (onsets, responses)
}
