% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodview_byblock.R
\name{json_foodview_byblock}
\alias{json_foodview_byblock}
\title{json_foodview_byblock: Generates json file for the derivative Food View Task data}
\usage{
json_foodview_byblock()
}
\value{
A JSON containing meta-data the summary Food View Task database
}
\description{
This function generates a json file for Food View summary data reported in derivatives/func-beh by block
}
