% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fsq.R
\name{json_fsq}
\alias{json_fsq}
\title{json_fsq: Generates a json file for the Feeding Strategies Questionnaire}
\usage{
json_fsq()
}
\value{
A json file documenting the raw inputs and scored values for the Feeding Strategies Questionnaire
}
\description{
This function generates a json file for the Feeding Strategies Questionnaire
This function is only accurate for data collected in Study REACH
}
