% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_hfi.R
\name{json_hfi}
\alias{json_hfi}
\title{json_hfi: Generates a json file for the Fulkerson Home Food Inventory}
\usage{
json_hfi()
}
\value{
A string with data stored in JSON format containing meta-data for the Fulkerson Home Food Inventory
}
\description{
This function generates a json file for the scored Fulkerson Home Food Inventory and raw participant responses.
}
