% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_hfias.R
\name{json_hfias}
\alias{json_hfias}
\title{json_hfias: Generates a json file for the Household Food Insecurity Access Scale}
\usage{
json_hfias()
}
\value{
A string with data stored in JSON format containing meta-data for the Household Food Insecurity Access Scale
}
\description{
This function generates a json file for the scored Binge Eating Scale and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_hfias function in dataprepr and is only accurate for data collected in Study REACH
}
