% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_household.R
\name{json_household}
\alias{json_household}
\title{json_household: Generates a json file for the household demographic questionnaire}
\usage{
json_household()
}
\value{
A string with data stored in JSON format containing meta-data for the household demographic questionnaire
}
\description{
This function provides a json file for household demographic questionnaire collected in Study REACH
}
