% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_infancy.R
\name{json_infancy}
\alias{json_infancy}
\title{json_infancy: Generates a json file for infancy data}
\usage{
json_infancy()
}
\value{
A string with data stored in JSON format containing meta-data for infant-related data
}
\description{
This function generates a json file for infant-related data collected in the visit 1 demographics form
}
