% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_kbas.R
\name{json_kbas}
\alias{json_kbas}
\title{json_kbas: Generates a json file for the kbas}
\usage{
json_kbas()
}
\value{
A string with data stored in JSON format containing meta-data for the KBAS
}
\description{
This function generates a json file for XXXX
}
