% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_liking.R
\name{json_liking}
\alias{json_liking}
\title{json_liking: Generates a json file for visits 1 and 5 VAS liking data}
\usage{
json_liking()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visits 1 and 5 VAS liking data
}
