% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_loc.R
\name{json_loc}
\alias{json_loc}
\title{json_loc: Generates a json file for the Loss of Control-Eating Questionnaire}
\usage{
json_loc()
}
\value{
A json file documenting the raw inputs and scored values for the Loss of Control-Eating Questionnaire.
}
\description{
This function generates a json file for the Loss of Control-Eating Questionnaire with raw participant responses. This function provides accurate json files ONLY if data was collected in Study REACH
}
