% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_mri_assessment.R
\name{json_mri_assessment}
\alias{json_mri_assessment}
\title{json_mri_assessment: Generates a json file for MRI visit liking and familiarity assessment}
\usage{
json_mri_assessment()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for MRI visit liking and familiarity assessment
}
