% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_mri_v2.R
\name{json_mri_v2}
\alias{json_mri_v2}
\title{json_mri_v2: Generates a json file for MRI visit 2 data}
\usage{
json_mri_v2()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for MRI visit 2 data
}
