% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_mri_vist.R
\name{json_mri_visit}
\alias{json_mri_visit}
\title{json_mri_visit: Generates a json file for MRI visit data}
\usage{
json_mri_visit()
}
\value{
A json file documenting the MRI visit data
}
\description{
This function generates a json file for MRI visit data (notes, CAMS, freddy fullness scores)
}
