% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_nihtoolbox_events.R
\name{json_nihtoolbox_events}
\alias{json_nihtoolbox_events}
\title{json_nihtoolbox_events: Generates a json file for the NIH toolbox events data (data in rawdata/beh)}
\usage{
json_nihtoolbox_events()
}
\value{
A string with data stored in JSON format containing meta-data for NIH toolbox assessment events (response) data
}
\description{
This function generates a json file for cleaned NIH toolbox assessment events (response) data
}
