% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_nihtoolbox_scores.R
\name{json_nihtoolbox_scores}
\alias{json_nihtoolbox_scores}
\title{json_nihtoolbox_scores: Generates a json file for the NIH toolbox behavioral response data (data in rawdata/beh)}
\usage{
json_nihtoolbox_scores()
}
\value{
A string with data stored in JSON format containing meta-data for NIH toolbox assessment (response) data
}
\description{
This function generates a json file for cleaned NIH toolbox assessment (response) data
}
