% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parent_updates.R
\name{json_parent_updates}
\alias{json_parent_updates}
\title{json_parent_updates: Generates a json file for parent_updates}
\usage{
json_parent_updates()
}
\value{
A string with data stored in JSON format containing meta-data for parent_updates
}
\description{
json_parent_updates: Generates a json file for parent_updates
}
