% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_participants.R
\name{json_participants}
\alias{json_participants}
\title{json_participants: Generates a json file for participants.tsv}
\usage{
json_participants()
}
\value{
A string with data stored in JSON format containing meta-data for participants.tsv
}
\description{
This function generates a json file for participants.tsv
}
