% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_phe_dataset_desc.R
\name{json_phe_dataset_desc}
\alias{json_phe_dataset_desc}
\title{json_phe_dataset_desc: Generate dataset_description JSON for phenotype}
\usage{
json_phe_dataset_desc(visit_data_path, data_de_path)
}
\arguments{
\item{visit_data_path}{full path to the redcap visit data used to generate phenotype data with proc_redcap.R}

\item{data_de_path}{full path to the redcap double entry data used to generate phenotype data with proc_redcap.R}
}
\value{
A string with data stored in JSON format containing meta-data for phenotype files derived by dataREACHr
}
\description{
This function generates a json file for phenotype/
}
