% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pit.R
\name{json_pit}
\alias{json_pit}
\title{json_pit: Generates a json file for the Pavlovian Instrumental Transfer task}
\usage{
json_pit()
}
\value{
A string with data stored in JSON format containing meta-data for the Pavlovian Instrumental Transfer task
}
\description{
This function generates a json file for cleaned Pavlovian Instrumental Transfer task
}
