% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pit_summary.R
\name{json_pit_summary}
\alias{json_pit_summary}
\title{json_pit_summary: Generates a json file for the Pavlovian Instrumental Transfer task summary data}
\usage{
json_pit_summary()
}
\value{
A string with data stored in JSON format containing meta-data for the Pavlovian Instrumental Transfer task summary data
}
\description{
This function generates a json file for cleaned Pavlovian Instrumental Transfer task summary data
}
