% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pmum.R
\name{json_pmum}
\alias{json_pmum}
\title{json_pmum: Generates a json file for the Problematic Media Use Measure}
\usage{
json_pmum()
}
\value{
A json file documenting the raw inputs and scored values for the Problematic Media Use Measure
}
\description{
This function generates a json file for the scored Problematic Media Use Measure and raw participant responses.
}
