% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pss.R
\name{json_pss}
\alias{json_pss}
\title{json_pss: Generates a json file for the 10-item Perceived Stress Scale}
\usage{
json_pss()
}
\value{
A json file documenting the raw inputs and scored values for the 10-item Perceived Stress Scale
}
\description{
This function generates a json file for the scored 10-item Perceived Stress Scale and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_pss function in dataprepr and is only accurate for data collected in Study REACH
}
