% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ptsca.R
\name{json_pstca}
\alias{json_pstca}
\title{json_pstca: Generates a json file for the XX}
\usage{
json_pstca()
}
\value{
A string with data stored in JSON format containing meta-data for the XX
}
\description{
This function generates a json file for the XX
}
