% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pstca.R
\name{json_ptsca}
\alias{json_ptsca}
\title{json_ptsca: Generates a json file for the XX}
\usage{
json_ptsca()
}
\value{
A string with data stored in JSON format containing meta-data for the XX
}
\description{
This function generates a json file for the XX
}
