% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_rank.R
\name{json_rank}
\alias{json_rank}
\title{json_rank: Generates a json file for the Ranking Food Item Questionnaire}
\usage{
json_rank()
}
\value{
A string with data stored in JSON format containing meta-data for the Ranking Food Item Questionnaire
}
\description{
This function generates a json file for the (scored?) Ranking Food Item Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_rank function in dataprepr and is only accurate for data collected in Study REACH
}
