% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_researcher_notes.R
\name{json_researcher_notes}
\alias{json_researcher_notes}
\title{json_researcher_notes: Generates a json file for researcher_notes}
\usage{
json_researcher_notes()
}
\value{
A string with data stored in JSON format containing meta-data for researcher_notes
}
\description{
json_researcher_notes: Generates a json file for researcher_notes
}
