% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_rrv.R
\name{json_rrv}
\alias{json_rrv}
\title{json_rrv: Generates a json file for RRV data in bids/rawdata}
\usage{
json_rrv()
}
\value{
A string with data stored in JSON format containing meta-data for the RRV task
}
\description{
This function generates a json file for RRV data in bids/rawdata
}
