% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_rrv_summary.R
\name{json_rrv_summary}
\alias{json_rrv_summary}
\title{json_rrv_summary: Generates a json file for the Relative Reinforcing Value of Food task summary data}
\usage{
json_rrv_summary()
}
\value{
A string with data stored in JSON format containing meta-data for the Relative Reinforcing Value of Food task summary data
}
\description{
This function generates a json file for cleaned Relative Reinforcing Value of Food task summary data
}
