% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_scpf.R
\name{json_scpf}
\alias{json_scpf}
\title{json_scpf: Generates a json file for the Structure and Control in Parent Feeding questionnaire}
\usage{
json_scpf()
}
\value{
A string with data stored in JSON format containing meta-data for the Structure and Control in Parent Feeding questionnaire
}
\description{
This function generates a json file for the scored Structure and Control in Parent Feeding questionnaire and raw participant responses.
}
