% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sic.R
\name{json_sic}
\alias{json_sic}
\title{json_sic: Generates a json file for the Stress in Children Questionnaire}
\usage{
json_sic()
}
\value{
A string with data stored in JSON format containing meta-data for the Stress in Children Questionnaire
}
\description{
This function generates a json file for the scored Stress in Children Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_sic function in dataprepr and is only accurate for data collected in Study REACH
}
