% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sleeplog.R
\name{json_sleeplog}
\alias{json_sleeplog}
\title{json_sleeplog: Generates a json file for the Sleep Log}
\usage{
json_sleeplog()
}
\value{
A string with data stored in JSON format containing meta-data for the Sleep Log
}
\description{
This function generates a json file for the Sleep Log
}
