% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_spacegame.R
\name{json_spacegame}
\alias{json_spacegame}
\title{json_spacegame: Generates json file for spacegame data (data in rawdata/beh)}
\usage{
json_spacegame()
}
\value{
A string with data stored in JSON format containing meta-data for the space game
}
\description{
json_spacegame: Generates json file for spacegame data (data in rawdata/beh)
}
