% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_spacegame_events.R
\name{json_spacegame_events}
\alias{json_spacegame_events}
\title{json_spacegame_events: Generates a json file for the Space Game events files}
\usage{
json_spacegame_events()
}
\value{
A json file documenting the individual level Space Game events files in rawdata
}
\description{
This function generates a json file for the individual level Space Game events files in rawdata
}
