% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_spsrq.R
\name{json_spsrq}
\alias{json_spsrq}
\title{json_spsrq: Generates a json file for the Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
\usage{
json_spsrq()
}
\value{
A json file documenting the raw inputs and scored values for the Sensitivity to Punishment and Sensitivity to Reward Questionnaire
}
\description{
This function generates a json file for the scored Sensitivity to Punishment and Sensitivity to Reward Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_spsrq function in dataprepr and is only accurate for data collected in Study REACH.
}
