% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sst_bold.R
\name{json_sst_bold}
\alias{json_sst_bold}
\title{json_sst_bold: Generates a json file for the SST fmri events (data in rawdata/func)}
\usage{
json_sst_bold()
}
\value{
A string with data stored in JSON format containing meta-data for the XXX
}
\description{
This function generates a json file for cleaned stop signal data collected during fmri (onsets, responses)
}
