% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sst_events.R
\name{json_sst_events}
\alias{json_sst_events}
\title{json_sst_events: Generates a json file for the SST fmri events (data in rawdata/func)}
\usage{
json_sst_events()
}
\value{
A json file documenting the XXX
}
\description{
This function generates a json file for cleaned stop signal data collected during fmri (onsets, responses)
}
