% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sst_prescan.R
\name{json_sst_prescan}
\alias{json_sst_prescan}
\title{json_sst_prescan: Generates a json file for the SST pre-scan behaviorald data (data in rawdata/beh)}
\usage{
json_sst_prescan()
}
\value{
A json file documenting the XXX
}
\description{
This function generates a json file for cleaned stop signal data collected during practice and behavioral task runs
}
