% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sst_summary.R
\name{json_sst_summary}
\alias{json_sst_summary}
\title{json_sst_summary: Generates json file for the derivative SST behavioral data}
\usage{
json_sst_summary()
}
\value{
a json file for SST summary data in wide format
}
\description{
This function generates a json file for SST summary data in wide format
}
