% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_stq.R
\name{json_stq}
\alias{json_stq}
\title{json_stq: Generates a json file for the Child Screen Time Questionnaire}
\usage{
json_stq()
}
\value{
A json file documenting the raw inputs and scored values for the Child Screen Time Questionnaire
}
\description{
This function generates a json file for the scored? Child Screen Time Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_stq function in dataprepr and is only accurate for data collected in Study REACH
}
