% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tfeq.R
\name{json_tfeq}
\alias{json_tfeq}
\title{json_tfeq: Generates a json file for the Three-Factor Eating Questionnaire-R18}
\usage{
json_tfeq()
}
\value{
A string with data stored in JSON format containing meta-data for the Three-Factor Eating Questionnaire-R18
}
\description{
This function generates a json file for the scored Three-Factor Eating Questionnaire-R18 and raw participant responses.
}
