% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_toolbox_beh.R
\name{json_toolbox_beh}
\alias{json_toolbox_beh}
\title{json_toolbox_beh: Generates a json file for the NIH toolbox behavioral response data (data in rawdata/beh)}
\usage{
json_toolbox_beh()
}
\value{
A string with data stored in JSON format containing meta-data for NIH toolbox assessment (response) data
}
\description{
This function generates a json file for cleaned NIH toolbox assessment (response) data
}
