% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_toolbox_phenotype.R
\name{json_toolbox_phenotype}
\alias{json_toolbox_phenotype}
\title{json_toolbox_phenotype: Generates a json file for the NIH toolbox behavioral score data (saved in data in bids/phentype)}
\usage{
json_toolbox_phenotype()
}
\value{
A string with data stored in JSON format containing meta-data for NIH toolbox score data
}
\description{
This function generates a json file for cleaned NIH toolbox score data
}
