% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_anthro.R
\name{json_v1_anthro}
\alias{json_v1_anthro}
\title{json_v1_anthro: Generates a json file for visit 1 anthropometrics data}
\usage{
json_v1_anthro()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for isit 1 anthropometrics data
}
