% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_data.R
\name{json_v1_data}
\alias{json_v1_data}
\title{json_v1_data: Generates a json file for general visit 1 notes and data}
\usage{
json_v1_data()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for ageneral visit 1 notes and data
}
