% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_freddy.R
\name{json_v1_freddy}
\alias{json_v1_freddy}
\title{json_v1_freddy: Generates a json file for visit 1 Freddy Fullness data}
\usage{
json_v1_freddy()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visit 1 Freddy Fullness data
}
