% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_infancy.R
\name{json_v1_infancy}
\alias{json_v1_infancy}
\title{json_v1_infancy: Generates a json file for visit 1 infancy data}
\usage{
json_v1_infancy()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visit 1 infancy data
}
