% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_intake.R
\name{json_v1_intake}
\alias{json_v1_intake}
\title{json_v1_intake: Generates a json file for intake data for V1}
\usage{
json_v1_intake()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for intake data for V1
}
