% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_intake_notde.R
\name{json_v1_intake_notde}
\alias{json_v1_intake_notde}
\title{json_v1_intake_notde: Generates a json file for intake data that has not been double-entered for V1}
\usage{
json_v1_intake_notde()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for intake data that has not been double-entered for V1
}
