% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v1_liking.R
\name{json_v1_liking}
\alias{json_v1_liking}
\title{json_v1_liking: Generates a json file for visit 1 VAS liking data}
\usage{
json_v1_liking()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visit 1 VAS liking data
}
