% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v3v4_intake.R
\name{json_v3v4_intake}
\alias{json_v3v4_intake}
\title{json_v3v4_intake: Generates a json file for intake data for visit 3 and visit 4}
\usage{
json_v3v4_intake()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for intake data for visit 3 and visit 4
}
