% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v3v4v5_freddy.R
\name{json_v3v4v5_freddy}
\alias{json_v3v4v5_freddy}
\title{json_v3v4v5_freddy: Generates a json file for visits 3-5 Freddy Fullness data}
\usage{
json_v3v4v5_freddy()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visits 3-5 Freddy Fullness data
}
