% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v3v4v5_intake.R
\name{json_v3v4v5_intake}
\alias{json_v3v4v5_intake}
\title{json_v3v4v5_intake: Generates a json file for intake data for visits 3-5}
\usage{
json_v3v4v5_intake()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for intake data for visits 3-5
}
