% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_actigraph.R
\name{proc_actigraph}
\alias{proc_actigraph}
\title{proc_actigraph: Process raw data actigraph data}
\usage{
proc_actigraph(base_wd, overwrite = FALSE, overwrite_ggir_derivs = FALSE)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{overwrite_ggir_derivs}{(logical) whether to re-process and overwrite all GGIR derivative outputs (this is a time intensive process). Default = FALSE.}
}
\value{
data.frame for each task with status for each processing step
}
\description{
This function calls task scripts that:
\itemize{
\item{1) processes raw actigraph data by session using GGIR (modes 1-5) with activity thresholds set according to: }
\itemize{
\item{Hildebrand M, VAN Hees VT, Hansen BH, Ekelund U. Age group comparability of raw accelerometer output from wrist- and hip-worn monitors. Med Sci Sports Exerc. 2014 Sep;46(9):1816-24. doi: 10.1249/MSS.0000000000000289. (\href{https://pubmed.ncbi.nlm.nih.gov/24887173/}{PubMed})}
\item{Hildebrand M, Hansen BH, van Hees VT, Ekelund U. Evaluation of raw acceleration sedentary thresholds in children and adults. Scand J Med Sci Sports. 2017 Dec;27(12):1814-1823. doi: 10.1111/sms.12795. Epub 2016 Nov 22. PMID: 27878845.(\href{https://pubmed.ncbi.nlm.nih.gov/27878845/}{PubMed})}
}
\item{completes post-processing with mMARCH.AC:}
\itemize{
\item{Guo, W., Leroux, A., Shou, H., Cui, L., Kang, S. J., Strippoli, M. F., Preisig, M., Zipunnikov, V., & Merikangas, K. R. (2023). Processing of Accelerometry Data with GGIR in Motor Activity Research Consortium for Health. Journal for the Measurement of Physical Behaviour, 6(1), 37-44. Retrieved Feb 14, 2025, from https://doi.org/10.1123/jmpb.2022-0018}
}
\item{computes Sleep Regularity Index scores using sleepreg:}
\itemize{
\item{Windred DP, Jones SE, Russell A, Burns AC, Chan P, Weedon MN, Rutter MK, Olivier P, Vetter C, Saxena R, Lane JM, Cain SW, Phillips AJK. Objective assessment of sleep regularity in 60 000 UK Biobank participants using an open-source package. Sleep. 2021 Dec 10;44(12):zsab254. doi: 10.1093/sleep/zsab254. PMID: 34748000.(\href{https://pubmed.ncbi.nlm.nih.gov/34748000/}{PubMed})}
\item{Windred DP, Burns AC, Lane JM, Saxena R, Rutter MK, Cain SW, Phillips AJK. Sleep regularity is a stronger predictor of mortality risk than sleep duration: A prospective cohort study. Sleep. 2024 Jan 11;47(1):zsad253. doi: 10.1093/sleep/zsad253. PMID: 37738616; PMCID: PMC10782501.(\href{https://pubmed.ncbi.nlm.nih.gov/37738616/}{PubMed})}
}
}
}
\details{
NOTE: to get GGIR to work with sleep log coded in ActiLife by trained RAs, d1_wakeup needed to be set to an arbitrary time (06:00:00). We do not have wakeup time for day 1 because actigraph was applied mid-day. Addid this time so we didn't loose night 1 sleep.
}
\examples{

# process task data for the Food Choice Task
proc_tasks_pardat <- proc_tasks(base_wd, overwrite)

\dontrun{
}


}
