% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_microstructure.R
\name{proc_microstructure}
\alias{proc_microstructure}
\title{proc_microstructure: Process raw microstructure coding data}
\usage{
proc_microstructure(base_wd, intake_data)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{intake_data}{processed intake data from proc_redcap.R}
}
\description{
This function loads the .txt raw data files from ObserverXT. Cleaning the data involves:
\itemize{
\item{1) Reads ObserverXT data from OneDrive}
\item{2) Calls util_ functions to clean and compile data in dataframes}
\item{3) Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe}
\item{4) Compiles data repeated across visits and sessions}
}
}
\examples{
#if in same working directory as data:
microstructure <- proc_microstructure(base_wd_path)

\dontrun{

}


}
