% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(redcap_api = FALSE, redcap_visit_data, redcap_de_data)
}
\arguments{
\item{redcap_api}{(logical) execute REDCap API. Default = FALSE.}

\item{redcap_visit_data}{REDCap visit data from a prior API call}

\item{redcap_de_data}{REDCap double-entry data from a prior API call}
}
\value{
Will return a list including:
\itemize{
\item{participants: list including data.frame with BIDS participants.tsv information and metadata}
\item{questionnaires: a list of 37 individual questionnaire lists including a data.frame and metadata}
\item{demographics}
\item{anthropometrics}
\item{intake}
\item{mri_visit}
\item{dexa}
\item{researcher_notes}
\item{parent_updates}
}
}
\description{
This function:
\itemize{
\item{1) Reads REDCap data (visit and double-entry) using the REDCap API}
\item{2) Calls util_ functions to clean and compile data in dataframes}
\item{3) Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe}
\item{4) Compiles data repeated across visits and sessions}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
redcap_data <- proc_redcap(base_wd, overwrite = FALSE, overwrite_jsons = FALSE)

}

}
