% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_tasks.R
\name{proc_tasks}
\alias{proc_tasks}
\title{proc_tasks: Process task data from untouchedRaw to create bids compliant files}
\usage{
proc_tasks(
  base_wd,
  overwrite = FALSE,
  all_tasks = FALSE,
  task_vector,
  overwrite_jsons = FALSE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) data should be overwritten. Default = FALSE}

\item{all_tasks}{(logical) all tasks should be moved to sourcedata. Default = FALSE. If all_tasks = FALSE, user must specify tasks to process in task_vector}

\item{task_vector}{(vector) tasks to process. Must be included if all_tasks = FALSE. Options include: c('sst', 'foodview', 'spacegame', 'nih_toolbox', 'rrv', 'pit') (optional)}

\item{overwrite_jsons}{(logical) overwrite existing jsons in rawdata. Default = FALSE}
}
\description{
This function: \itemize{
\item{1) copies task data from untouchedRaw into bids/sourcedata for all tasks (food view, pit, sst, nih toolbox, spacegame), using util_task_untouched_to_source(all_tasks = TRUE)}
\item{2) processes task sourcedata and exports cleaned dataframes into bids/rawdata for the following tasks: rrv, sst, foodview nih-toolbox, pit, spacegame, using task-specific util_task_{task-name} functions}
\item{3) exports JSON meta-data files for tasks organized into rawdata (rrv, sst, foodview), using write_task_jsons()}
}
}
\examples{

\dontrun{
# organize task data for all tasks in untouchedRaw into sourcedata and rawdata
proc_task(base_wd = base_wd, all_tasks = TRUE)

# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
proc_task(base_wd = base_wd, task_vector = c('spacegame', 'nih_toolbox'))

}


}
