% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_redcap.R
\name{qc_redcap}
\alias{qc_redcap}
\title{qc_redcap: Check processed redcap data for quality issues}
\usage{
qc_redcap(phenotype_data)
}
\arguments{
\item{phenotype_data}{a list of phenotype dataframes generated by proc_redcap()}
}
\description{
This function checks processed redcap data for the following qualities:
\itemize{
\item{Duplicate data:}
\itemize{
\item{Checks participant dataframe for duplicate subject IDs}
\item{Checks intake dataframe for duplicate subject + visit combinations}
\item{Checks puberty dataframe for duplicate subject + session + respondent combinations}
\item{Checks all other dataframes for duplicate subject and session combinations}
}
\item{Data types:}
\itemize{
\item{Checks that intake data (pre and post-weights) are numerical}
\item{Checks LOC data: }
}
\item{Data values:}
\itemize{
\item{Checks child ages at V1 between >8 and <10}
\item{Checks intake consumption (g) <= amount served}
}
\item{Missing data:}
\itemize{
\item{}
}
}
}
\examples{

\dontrun{

# return data from proc_redcap()
redcap_data <- proc_redcap(visit_data_path, data_de_path, return = TRUE)
phenotype_data <- redcap_data$phenotype_data

# run QC on phenotype_data in redcap_data
qc_redcap(phenotype_data)

}
}
