% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_export.R
\name{redcap_export}
\alias{redcap_export}
\title{redcap_export:}
\usage{
redcap_export(data_str, overwrite)
}
\description{
redcap_export:
}
\examples{

# process data
ed_data <- util_gen_ed_data(intake_data)


}
