% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrv_parse_text.R
\name{rrv_parse_text}
\alias{rrv_parse_text}
\title{rrv_parse_text: Parse RRV txt files when csv not available}
\usage{
rrv_parse_text(sub, rrv_file, overwrite = FALSE, return_data = TRUE)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{rrv_file}{string with full path to rrv file}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return XXX
}
\description{
This function formats and organizes Food View task data from bids/sourcedata into bids/rawdata for a given subject
}
\examples{

\dontrun{

# process task data for the Food View Task
file_name = "~/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/untouchedRaw/rrv_task/060.txt"
rrv_csv <- rrv_parse_text(sub = 60, rrv_file = file_name)

}


}
