% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_actigraph_clean.R
\name{util_actigraph_clean}
\alias{util_actigraph_clean}
\title{util_actigraph_clean: Compile and clean up processed actigraphy data}
\usage{
util_actigraph_clean(
  base_wd,
  metrics = c("SL", "PA", "CR"),
  overwrite = FALSE,
  return_data = TRUE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{metrics}{list of metrics. Default: c('SL', 'PA', 'CR'). Options include: \itemize{
\item{SL - sleep}
\item{PA - physical activity}
\item{CR - circadian rhythem}
}}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{return_data}{(logical) return data to working environment. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function compiles and cleans up processed actigraphy data from GGIR and mMARCH.AC
}
\details{
To use this function, util_actigraph_org_rawdata.R and util_actigraph_mMARCH.R must be completed so all actigraphy information is available.
}
\examples{

# process task data for the Food Choice Task
sleep_dat <- util_actigraph_clean(base_wd, metrics = c('SL'), return = TRUE)

\dontrun{
}


}
