% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_actigraph_mMARCH.R
\name{util_actigraph_mMARCH}
\alias{util_actigraph_mMARCH}
\title{util_actigraph_mMARCH: Set up and run mMARCH.AC.maincall() from the mMARCH.AC package}
\usage{
util_actigraph_mMARCH(
  mode,
  deriv_dir,
  study_name,
  data_list,
  ggir_path,
  filename2id
)
}
\arguments{
\item{mode}{the mode to run in mMARCH.AC. Notes: \itemize{
\item{0 = create R/Rmd/sh files}
\item{1 = read, transform, and merg data from csv folder of GGIR}
\item{2 = GGIR output files checked and summarized}
\item{3 = clean merged data}
\item{4 = imputation of cleaned data}
}}

\item{deriv_dir}{full path to the derivative directory for output to be stored}

\item{study_name}{character string with study name}

\item{data_list}{list of file paths to raw data}

\item{ggir_path}{full path to the GGIR output files}

\item{filename2id}{a function defined to extract subject id from filenames}
}
\description{
This function fills out necessary settings to set up and run mMARCH.AC post GGIR processing.
}
\details{
Notes:
\itemize{
\item{1) Uses activity thresholds set according to: }
\itemize{
\item{Hildebrand M, VAN Hees VT, Hansen BH, Ekelund U. Age group comparability of raw accelerometer output from wrist- and hip-worn monitors. Med Sci Sports Exerc. 2014 Sep;46(9):1816-24. doi: 10.1249/MSS.0000000000000289. (\href{https://pubmed.ncbi.nlm.nih.gov/24887173/}{PubMed})}
\item{Hildebrand M, Hansen BH, van Hees VT, Ekelund U. Evaluation of raw acceleration sedentary thresholds in children and adults. Scand J Med Sci Sports. 2017 Dec;27(12):1814-1823. doi: 10.1111/sms.12795. Epub 2016 Nov 22. PMID: 27878845.(\href{https://pubmed.ncbi.nlm.nih.gov/27878845/}{PubMed})}
}
\item{2) Relies on GGIR outpus so must be completed after GGIR pre-processing}
}

References:
\itemize{
\item{Guo, Wei, Andrew Leroux, Haochang Shou, Lihong Cui, Sun Jung Kang, Marie Pierre Françoise Strippoli, Martin Preisig, Vadim Zipunnikov, and Kathleen Ries Merikangas. "Processing of Accelerometry Data with GGIR in Motor Activity Research Consortium for Health." Journal for the Measurement of Physical Behaviour 6, no. 1 (2023): 37-44.}
}

To use this function, the correct paths for all directories must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# setup post-processing for actigraph data
util_actigraph_mMARCH(deriv_dir, sub_str, study_name, data_list, ggir_path)

\dontrun{
}


}
