% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_actigraph_org_sourcedata.R
\name{util_actigraph_org_sourcedata}
\alias{util_actigraph_org_sourcedata}
\title{util_actigraph_org_sourcedata: Organize actigraphy raw data into BIDS format sourcedata}
\usage{
util_actigraph_org_sourcedata(base_wd, sub_str, ses_str, overwrite = FALSE)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}
}
\description{
This function copies data and saves it in sourcedata
}
\examples{

# organize actigraphy data
org_actigraph <- util_actigraph_org_sourcedata(tsub_str = 'sub_001', ses_str = 'ses-1', base_wd, overwrite = TRUE)

\dontrun{
}


}
