% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_de_check.R
\name{util_de_check}
\alias{util_de_check}
\title{util_de_check: Checks double-entry data that have not yet been merged in REDCap}
\usage{
util_de_check(unmerged_data)
}
\arguments{
\item{unmerged_data}{data.frame of unmerged double entry data (should have ids similar to 001--1, 001--2)}
}
\value{
Two data.frames:
\itemize{
\item{1) merged_de_data: merged data.frame after confirming matching values}
\item{2) unmerged_de_data: data.frame containing participants with unmatched or non-double-entered data}
\item{3) single_de_data: data.frame contianing participant with only 1 set of entered values so no comparison could be made}
}
}
\description{
This function will compare double-entry data that have not yet merged in REDCap
}
\examples{

# process data
data_de_list <- util_de_check(unmerged_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
