% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_foodview_summary.R
\name{util_foodview_summary}
\alias{util_foodview_summary}
\title{util_foodview_summary: Get summary data from individual participants for the Food View task}
\usage{
util_foodview_summary(ind_dat, format = "wide")
}
\arguments{
\item{ind_dat}{Processed individual dataset from rawdata for the Food View task}

\item{format}{data format to return. 'wide' will return summary statistics by commercial condition (food vs toy), 'long' will return data by run and block. Defualt = 'wide'}
}
\value{
a data.frame with summary performance and task metrics for a participant
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
foodview_summary_beh <- util_foodview_summary(ind_dat)

\dontrun{
}


}
