% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_actigraph_sleeplog.R
\name{util_format_actigraph_sleeplog}
\alias{util_format_actigraph_sleeplog}
\title{util_format_actigraph_sleeplog: formats actigraph scored sleep data for GGIR}
\usage{
util_format_actigraph_sleeplog(sleeplog_dat)
}
\arguments{
\item{sleeplog_dat}{scored sleeplog data from actigraph watches using ActiLife}
}
\description{
This function prepares actigraph scored sleep data for GGIR processing
}
\examples{

# process data
sleeplog_dat_formatted <- util_format_actigraph_sleeplog(sleeplog_dat)

}
\seealso{
\code{\link[=proc_actigraph]{proc_actigraph()}}
}
