% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_cshq_data.R
\name{util_format_cshq_data}
\alias{util_format_cshq_data}
\title{util_format_cshq_data: process cshq data (called within util_redcap_parent_v2 and util_redcap_parent_v5)}
\usage{
util_format_cshq_data(cshq_data)
}
\arguments{
\item{cshq_data}{cshq extracted from data from REDCap events util_redcap_parent_v2 and util_redcap_parent_v5}
}
\description{
This function process cshq data
}
