% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_fhfi_data.R
\name{util_format_fhfi_data}
\alias{util_format_fhfi_data}
\title{util_format_fhfi_data: Prepare Fulkerson HFI data for scoring (called within util_redcap_parent_v4.R)}
\usage{
util_format_fhfi_data(data)
}
\arguments{
\item{data}{Fulkerson HFI extracted from data from REDCap event parent_visit_4_arm_1'}
}
\description{
This function prepares Fulkerson HFI data for scoring with dataprepr::score_hfi()
}
