% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_hfi_data.R
\name{util_format_hfi_data}
\alias{util_format_hfi_data}
\title{util_format_hfi_data: Prepare Fulkerson HFI data for scoring}
\usage{
util_format_hfi_data(hfi_data)
}
\arguments{
\item{hfi_data}{Fulkerson HFI extracted from data from REDCap event}
}
\description{
This function prepares Fulkerson HFI data for scoring with dataprepr::score_hfi()
}
\examples{

# process data
hfi_data_formatted <- util_format_hfi_data(hfi_data)

}
\seealso{
\code{\link[=util_redcap_parent_v4]{util_redcap_parent_v4()}}
}
