% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_kbas_data.R
\name{util_format_kbas_data}
\alias{util_format_kbas_data}
\title{util_format_kbas_data: process kbas data (called within util_redcap_child_v1 and util_redcap_child_v5)}
\usage{
util_format_kbas_data(kbas_data)
}
\arguments{
\item{kbas_data}{kbas extracted from data from REDCap events util_redcap_child_v1 and util_redcap_child_v5}
}
\description{
This function process kbas data
}
