% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_sleep_data.R
\name{util_format_sleep_data}
\alias{util_format_sleep_data}
\title{util_format_sleep_data: format Sleep Log data}
\usage{
util_format_sleep_data(sleeplog_data)
}
\arguments{
\item{sleeplog_data}{Sleep Log data from util_redcap_child_v3}
}
\value{
formatted sleeplog_data
}
\description{
This function formats Parent-reported sleep log data
}
\examples{

# process kbas data
child_sleeplog <- util_format_sleep_data(data)

}
\seealso{
\code{\link[=util_redcap_child_v3]{util_redcap_child_v3()}}
}
