% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_gen_ed_data.R
\name{util_gen_ed_data}
\alias{util_gen_ed_data}
\title{util_gen_ed_data: Generate a dataframe with energy density data}
\usage{
util_gen_ed_data()
}
\description{
This function generates a dataframe with energy density data for processing intake data with util_redcap_de()
}
