% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_group_foodview.R
\name{util_group_foodview}
\alias{util_group_foodview}
\title{util_group_foodview: Get summary data from the Food Rating task}
\usage{
util_group_foodview(
  data_list,
  ses_str,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{data_list}{A data frame with variable 'sub_str' that includes all participants that have task data in rawdata}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{return}{logical indicating if computed summary data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function calculates summary performance data and saves the output
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Rating Task
group_foodview_data <- util_group_foodview(data_list, ses, base_wd, return = TRUE)

\dontrun{
}


}
