% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_group_spacegame.R
\name{util_group_spacegame}
\alias{util_group_spacegame}
\title{util_group_spacegame: Get summary data from the Space Game (2-stage reinforcement learning task)}
\usage{
util_group_spacegame(
  data_list,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{return}{logical indicating if computed summary data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function calculates summary performance data and saves the output
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.

#' @param data_list A data frame with variable 'sub_str' that includes all participants that have task data in rawdata
}
\examples{

# process task data for the Food Choice Task
group_spacegame_dat <- util_group_spacegame(ddata_list, ses, base_wd, overwrite, return = TRUE)

\dontrun{
}


}
