% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_demo.R
\name{util_merge_demo}
\alias{util_merge_demo}
\title{util_merged_demo: merges and formats demographic data}
\usage{
util_merge_demo(visit1_demo, household_all, merged_anthro, date_data)
}
\arguments{
\item{visit1_demo}{visit 1 demo data.frame}

\item{household_all}{merged household data.frame}

\item{merged_anthro}{merged anthro data.fram}

\item{date_data}{date data for REDCap visit

Will return a list including:
\itemize{
\item{clean raw and scored parent visit 1 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{demo_data}
\item{infancy_data}
\item{household_data}
\item{rank_data}
\item{puberty_data}
\item{cfq_data}
\item{cebq_data}
\item{efcr_data}
\item{chaos_data}
\item{pss_data}
\item{lbc_data}
}}
}
\description{
This function merges demographic data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_demo_data <- util_merged_demo(parent_v1_data$demo_data$data, household_all, merged_anthro, date_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_merge_anthro]{util_merge_anthro()}}, , \code{\link[=util_merge_questionnaires]{util_merge_questionnaires()}}
}
