% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_questionnaires.R
\name{util_merge_questionnaires}
\alias{util_merge_questionnaires}
\title{util_merge_questionnaires: merges data for questionnaires that were repeated at baseline and follow-up}
\usage{
util_merge_questionnaires(
  child_v1_data,
  child_v4_data,
  child_v5_data,
  parent_v1_data,
  parent_v2_data,
  parent_v3_data,
  parent_v4_data,
  parent_v5_data
)
}
\arguments{
\item{child_v1_data}{full set of child visit 1 data generated by util_redcap_child_v1}

\item{child_v4_data}{full set of child visit 5 data generated by util_redcap_child_v4}

\item{child_v5_data}{full set of child visit 5 data generated by util_redcap_child_v5}

\item{parent_v1_data}{full set of parent visit 1 data generated by util_redcap_parent_v1}

\item{parent_v2_data}{full set of parent visit 3 data generated by util_redcap_parent_v2}

\item{parent_v3_data}{full set of parent visit 3 data generated by util_redcap_parent_v3}

\item{parent_v4_data}{full set of parent visit 4 data generated by util_redcap_parent_v4}

\item{parent_v5_data}{full set of parent visit 5 data generated by util_redcap_parent_v5}
}
\value{
Will return a list of merged data files:
\itemize{
\item{stq_all}
\item{kbas_all}
\item{loc_all}
\item{household_all}
\item{cebq_all}
\item{cbq_all}
\item{cshq_all}
\item{pstca_all}
\item{audit_all}
\item{pmum_all}
\item{cfpq_all}
\item{rank_all}
\item{class_all}
\item{puberty_all}
\item{updates_all}
}
}
\description{
This function data for questionnaires that were repeated at baseline and follow-up. The following questionnaires will be merged across visits and returned individually in a list:
}
\examples{

# process data
merged_qs_list <- util_merge_questionnaires(child_v1_data, child_v4_data, child_v5_data, parent_v1_data, parent_v2_data, parent_v3_data, parent_v4_data, parent_v5_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_redcap_child_v1]{util_redcap_child_v1()}}, \code{\link[=util_redcap_child_v4]{util_redcap_child_v4()}}, \code{\link[=util_redcap_child_v5]{util_redcap_child_v5()}}, \code{\link[=util_redcap_parent_v1]{util_redcap_parent_v1()}}, \code{\link[=util_redcap_parent_v2]{util_redcap_parent_v2()}}, \code{\link[=util_redcap_parent_v3]{util_redcap_parent_v3()}}, \code{\link[=util_redcap_parent_v4]{util_redcap_parent_v4()}}, \code{\link[=util_redcap_parent_v5]{util_redcap_parent_v5()}}
}
