% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_anthro.R
\name{util_merged_anthro}
\alias{util_merged_anthro}
\title{util_merged_anthro: merges and formats anthropometrics data}
\usage{
util_merged_anthro(visit1_anthro, visit5_anthro, merged_household, date_data)
}
\arguments{
\item{visit1_anthro}{visit 1 anthropometrics data.frame}

\item{visit5_anthro}{visit 2 anthropometrics data.frame}

\item{merged_household}{merged household data.frame}

\item{date_data}{date data for REDCap visit

Will return a list including:
\itemize{
\item{clean raw and scored parent visit 1 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{demo_data}
\item{infancy_data}
\item{household_data}
\item{rank_data}
\item{puberty_data}
\item{cfq_data}
\item{cebq_data}
\item{efcr_data}
\item{chaos_data}
\item{pss_data}
\item{lbc_data}
}}
}
\description{
This function merges anthropometrics data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_anthro_data <- util_merged_anthro(child_v1_data$anthro_data$data, child_v5_data$anthro_data$data, merged_qs_list$household_all, date_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_merge_questionnaires]{util_merge_questionnaires()}}
}
