% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_demo.R
\name{util_merged_demo}
\alias{util_merged_demo}
\title{util_merged_demo: merges and formats demographic data}
\usage{
util_merged_demo(visit1_demo, household_all, merged_anthro, date_data)
}
\arguments{
\item{visit1_demo}{visit 1 demo data.frame}

\item{household_all}{merged household data.frame}

\item{merged_anthro}{merged anthro data.fram}

\item{date_data}{data.frame with all visit date information}
}
\description{
This function merges demographic data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_demo_data <- util_merged_demo(parent_v1_data$demo_data$data, household_all, merged_anthro, date_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_merge_anthro]{util_merge_anthro()}}, , \code{\link[=util_merge_questionnaires]{util_merge_questionnaires()}}
}
