% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_intake.R
\name{util_merged_intake}
\alias{util_merged_intake}
\title{util_merged_intake: merges and formats all data related to meals and EAH}
\usage{
util_merged_intake(child_v1_data, child_v3_data, child_v4_data, child_v5_data)
}
\arguments{
\item{child_v1_data}{full set of child visit 1 data generated by util_redcap_child_v1}

\item{child_v3_data}{full set of child visit 3 data generated by util_redcap_child_v3}

\item{child_v4_data}{full set of child visit 4 data generated by util_redcap_child_v4}

\item{child_v5_data}{full set of child visit 5 data generated by util_redcap_child_v5}
}
\description{
This function merges all data related to meals and EAH and calculates intake values
}
\examples{

# process data
merged_intake <- util_merged_intake(child_v1_data, child_v3_data, child_v4_data, child_v5_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_redcap_child_v1]{util_redcap_child_v1()}}, \code{\link[=util_redcap_child_v3]{util_redcap_child_v3()}}, \code{\link[=util_redcap_child_v4]{util_redcap_child_v4()}}, \code{\link[=util_redcap_child_v5]{util_redcap_child_v5()}}, \code{\link[=util_calc_intake]{util_calc_intake()}}
}
