% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_participants.R
\name{util_merged_participants}
\alias{util_merged_participants}
\title{util_merged_participants: merges and formats participants data for BIDS}
\usage{
util_merged_participants(visit1_demo, merged_demo, date_data)
}
\arguments{
\item{visit1_demo}{visit 1 demo data.frame}

\item{merged_demo}{merged demo data.frame}

\item{date_data}{data.frame with all visit date information}
}
\description{
This function merges demographic data across visits and formats/calculates necessary values to make the participants.tsv BIDS file
}
\examples{

# process data
merged_participants <- util_merged_participants(parent_v1_data$demo_data$data, merged_demo, date_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_merge_demo]{util_merge_demo()}}
}
