% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_visitnotes.R
\name{util_merged_visitnotes}
\alias{util_merged_visitnotes}
\title{util_merged_visitnotes: merges and formates visit notes}
\usage{
util_merged_visitnotes(
  visit1_notes,
  visit2_notes,
  visit3_notes,
  visit4_notes,
  visit5_notes
)
}
\arguments{
\item{visit1_notes}{visit 1 notes data.frame}

\item{visit2_notes}{visit 2 notes data.frame}

\item{visit3_notes}{visit 3 notes data.frame}

\item{visit4_notes}{visit 4 notes data.frame}

\item{visit5_notes}{visit 5 notes data.frame}
}
\description{
This function merges visit notes in wide format and re-names variables
}
\examples{

# process data
visit_notes_data <- util_merged_visitnotes(visit1_notes, visit2_notes, visit3_notes, visit4_notes, visit5_notes)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
