% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_micro_switch.R
\name{util_micro_switch}
\alias{util_micro_switch}
\title{util_micro_switch: Code switching between microstructure behaviors}
\usage{
util_micro_switch(id_var, micro_data, switch_var, coder_var)
}
\arguments{
\item{id_var}{(string) name of participant ID variable}

\item{micro_data}{data frame that contains ObserverXT event data where column id_var has values matching id}

\item{switch_var}{(string) name of variable in micro_data to code switching on - must match variable name in micro_data}

\item{coder_var}{(string) name of variable that indicates coder order (e.g., 1, 2 or 'coder 1', 'coder 2') if some participants were coded by multiple coders. micro_data should be in long format with coded events for each coder stacked rather than organized in separate columns}
}
\value{
Will return a data.frame with the following variable:
\itemize{
\item{switch_var: coded switches}
}
}
\description{
This function code switches between microstructure behaviors for all participants in micro_data
}
\examples{
#if in same working directory as data:
switching <- util_micro_switch('participant_id', micro_data, 'food', 'coder_order')

\dontrun{

}


}
