% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_micro_wide.R
\name{util_micro_wide}
\alias{util_micro_wide}
\title{util_micro_wide: Separate microstructure summary behavior by coder for a single participant}
\usage{
util_micro_wide(id_var, micro_data, ses_num, paradigm, coder_var)
}
\arguments{
\item{id_var}{(string) name of participant ID variable}

\item{micro_data}{data frame that contains ObserverXT event data where column id_var has values matching id}

\item{ses_num}{session number (option: 1 or 2)}

\item{paradigm}{eating paradigm entered as string. Options include: 'meal' or 'eah'}

\item{coder_var}{(string) name of variable that indicates coder order (e.g., 1, 2 or 'coder 1', 'coder 2') if some participants were coded by multiple coders. micro_data should be in long format with coded events for each coder stacked rather than organized in separate columns}
}
\value{
a data frame with 1 row and with data by coder for the following variables:
\itemize{
\item{nbites}
\item{nsips}
\item{meal_dur}
\item{bite_latency}
\item{total_active_eating}
\item{nbeh_distract}
\item{start_meal_affect}
\item{end_meal_affect}
\item{nswitch}
\item{nswitch_ed}
\item{nswitch_wsip}
\item{nswitch_wsip_distract}
\item{bite_rate}
\item{bite_rate_active}
\item{sip_rate}
\item{sip_rate_active}
\item{switch_rate}
\item{switch_ed_rate}
\item{switch_wsip_rate}
}
}
\description{
This function extract microstructure summary behavior by coder
}
\examples{
#if in same working directory as data:
beh_wide <- util_micro_wide('participant_id', micro_data, ses_num = 1, paradigm = 'meal', coder_var
'coder_order')

\dontrun{

}


}
