% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_org_sourcedata.R
\name{util_org_sourcedata}
\alias{util_org_sourcedata}
\title{util_org_sourcedata: Move task data from untouchedRaw into bids/sourcedata}
\usage{
util_org_sourcedata(base_wd, overwrite = FALSE)
}
\arguments{
\item{base_wd}{string with full path to base directory -- this is the directory that contains untouchedraw/ and bids/sourcedata/}

\item{overwrite}{logical indicating if data should be overwritten in /sourcedata Default = FALSE}
}
\description{
This function moves task data (food view, sst) from untouchedRaw into bids/sourcedata
}
\examples{

# organize task data in untouchedRaw into sourcedata
util_org_sourcedata(base_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/")

\dontrun{
}


}
