% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_phenotype_toolbox.R
\name{util_phenotype_toolbox}
\alias{util_phenotype_toolbox}
\title{util_phenotype_toolbox: Clean and organize NIH toolbox score data into phenotype/toolbox.tsv}
\usage{
util_phenotype_toolbox(
  sub,
  ses,
  bids_wd,
  overwrite = FALSE,
  return_data = TRUE
)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required (integer)}

\item{ses}{session label used in ses-label (integer)}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a dataframe with NIH toolbox score variables for given subject
}
\description{
This function formats and adds NIH toolbox scored data from bids/sourcedata into phenotype/toolbox.tsv for a given subject
}
\examples{

\dontrun{
# process score data for the NIH toolbox
sub001_toolbox_scores <- util_phenotype_toolbox(sub = 001, ses = 1, bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}

}
