% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pit_summary.R
\name{util_pit_summary}
\alias{util_pit_summary}
\title{util_pit_summary: Get summary data from individual participants for the Pavlovian Instramental Transfer task}
\usage{
util_pit_summary(ind_dat)
}
\arguments{
\item{ind_dat}{Processed individual dataset from rawdata for the Pavlovian Instramental Transfer task}
}
\value{
a data.frame with 3 rows including summary performance and task metrics by response type for a participant
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
pit_summary_beh <- util_pit_summary(ind_dat)

\dontrun{
}


}
