% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v1.R
\name{util_redcap_child_v1}
\alias{util_redcap_child_v1}
\title{util_redcap_child_v1: Organize child visit 1 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_child_v1(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event child_visit_1_arm_1}

\item{return_data}{If return_data is set to TRUE, will return a list of dataframes including: visit_data_child, food_paradigm_info, freddy_data, intake_data, liking_data, kbas_data, stq_data, anthro_data}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_1_arm_1
}
