% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v2.R
\name{util_redcap_child_v2}
\alias{util_redcap_child_v2}
\title{util_redcap_child_v2: Organize child visit 2 data from REDCap}
\usage{
util_redcap_child_v2(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event child_visit_2_arm_1}
}
\value{
Will return a list including:
\itemize{
\item{clean raw child visit 1 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{visit_data_child}
\item{mri_info}
\item{mri_assessment_data}
}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_2_arm_1
}
\examples{

# process REDCap data
child_visit2_list <- util_redcap_child_v2(data)

\dontrun{
}

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
