% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v3.R
\name{util_redcap_child_v3}
\alias{util_redcap_child_v3}
\title{util_redcap_child_v3: Organize child visits 3 data from REDCap}
\usage{
util_redcap_child_v3(data)
}
\arguments{
\item{data}{data from REDCap event child_visit_3_arm_1}
}
\value{
Will return a list including:
\itemize{
\item{clean raw child visit 3 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{visit_data_child}
\item{food_paradigm_info}
\item{freddy_data}
\item{intake_data}
\item{eah_wanting}
\item{sleeplog_data}
}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_3_arm_1
}
\examples{

# process REDCap data
child_visit3_list <- util_redcap_child_v3(data)

\dontrun{
}

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
