% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_dates.R
\name{util_redcap_dates}
\alias{util_redcap_dates}
\title{util_redcap_dates: Organize child date data from REDCap across all visits (called within proc_redcap.R)}
\usage{
util_redcap_dates(child_v1, child_v2, child_v3, child_v4, child_v5, parent_v1)
}
\arguments{
\item{child_v1}{data from REDCap child visit 1 event}

\item{child_v2}{data from REDCap child visit 2 event}

\item{child_v3}{data from REDCap child visit 3 event}

\item{child_v4}{data from REDCap child visit 4 event}

\item{child_v5}{data from REDCap child visit 5 event}

\item{parent_v1}{data from REDCap parent visit 1 event}
}
\description{
This function organizes date data from REDCap across visits and events
}
