% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rrv_parse_text.R
\name{util_rrv_parse_text}
\alias{util_rrv_parse_text}
\title{util_rrv_parse_text: Generate tabular data from rrv text files}
\usage{
util_rrv_parse_text(rrv_file, sub_str)
}
\arguments{
\item{rrv_file}{(character string) full path to rrv file}

\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}
}
\description{
This function generates and returns 1 dataframe with data from input RRV file
}
\examples{

\dontrun{

# process task data for the RRV Task
rrv_csv <- util_rrv_parse_text(rrv_file = file_name, sub_str = 'sub-060')

}


}
