% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rrv_summary.R
\name{util_rrv_summary}
\alias{util_rrv_summary}
\title{util_rrv_summary: Get summary data from individual participants for tthe Reinforcing Value of Food task}
\usage{
util_rrv_summary(ind_dat, format)
}
\arguments{
\item{ind_dat}{Processed individual dataset from rawdata for the the Reinforcing Value of Food task task}

\item{format}{Format in which the summary data is returned: 'wide' will return a 1-row summary data.frame(); 'long' will return the summary session response data by session and screen.}
}
\value{
a data.frame with summary data based on format requested
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
rrv_summary_beh <- util_rrv_summary(ind_dat)

\dontrun{
}


}
