% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_spacegame_summary.R
\name{util_spacegame_summary}
\alias{util_spacegame_summary}
\title{util_spacegame_summary: Get summary data from from individual participants for Space Game (2-stage reinforcement learning task)}
\usage{
util_spacegame_summary(ind_data)
}
\arguments{
\item{ind_data}{Processed individual dataset from rawdata for the Space Game task}
}
\value{
a data.frame with 1 row including summary performance and task metrics for a participant
}
\description{
This function calculates summary performance from individual participant data
}
\examples{

# process task data for the Food Choice Task
spacegame_summary_dat <- util_spacegame_summary(ddata_list, ses, base_wd, overwrite, return = TRUE)

\dontrun{
}


}
