% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sst_summary.R
\name{util_sst_summary}
\alias{util_sst_summary}
\title{util_sst_summary: Get summary data from individual participants for the Stop Signal Task}
\usage{
util_sst_summary(ind_dat, format)
}
\arguments{
\item{ind_dat}{Processed individual dataset from rawdata for the Stop Signal Task task}

\item{format}{Format in which the summary data is returned: 'wide' will return a 3-row summary data.frame() for all trials, behavioral pre-scan trials, and fMRI trials; 'byrun' will return the summary response data by run; 'byblock' will return the summary response data by run and block.}
}
\value{
a data.frame with summary data based on format requested
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
sst_summary_beh <- util_sst_summary(ind_dat)

\dontrun{
}


}
