% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_foodview.R
\name{util_task_foodview}
\alias{util_task_foodview}
\title{util_task_foodview: Clean and organize Food View task data into BIDS rawdata}
\usage{
util_task_foodview(sub_str, ses_str = "ses-1", bids_wd, overwrite = FALSE)
}
\arguments{
\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{bids_wd}{string with full path to bids directory that contains the sourcedata/ and rawdata/ directories}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list with 1 cleaned dataframe per run
}
\description{
This function formats and organizes Food View task data from bids/sourcedata into bids/rawdata for a given subject
}
\examples{

\dontrun{
# process task data for the Food View Task
util_task_foodview_orgraw(sub_id = 'sub-001', ses = 1, bids_wd = bids_wd)

}


}
