% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_nihtoolbox.R
\name{util_task_nihtoolbox}
\alias{util_task_nihtoolbox}
\title{util_task_nihtoolbox: Clean and organize NIH toolbox assessment data into rawdata}
\usage{
util_task_nihtoolbox(sub_str, ses_str = "ses-1", bids_wd, overwrite = FALSE)
}
\arguments{
\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{bids_wd}{string with full path to bids directory that contains the sourcedata/ and rawdata/ directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}
}
\value{
statement of task completed
}
\description{
This function formats and organizes NIH toolbox assessment data from bids/sourcedata into rawdata for a given subject. Assessment data includes responses to each trial in the NIH toolbox.
}
\examples{

\dontrun{
# process assessment (response) data for the NIH toolbox
util_task_nihtoolbox(sub_str = 'sub-001', ses_str = 'ses-1', bids_wd = bids_wd)

}

}
