% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_rrv.R
\name{util_task_rrv}
\alias{util_task_rrv}
\title{util_task_rrv: Clean and organize RRV data into BIDS rawdata}
\usage{
util_task_rrv(sub_str, ses_str, bids_wd, overwrite = FALSE, return_data = TRUE)
}
\arguments{
\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{bids_wd}{string with full path to bids directory that contains the sourcedata/ and rawdata/ directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}
}
\value{
statement of task completed
}
\description{
This function formats and organizes RRV data from bids/sourcedata into bids/rawdata for a given subject
If subject has RRV text file, rrv_parse_text() will be called. If subject does not have text file but has game.csv, game.csv will processed to generate a dataframe like those generated by rrv_parse_text()
}
\examples{

\dontrun{
# process task data for the Food View Task
rrv_data <- util_task_rrv(sub_str = 'sub-001', ses_str = 'ses-1', bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}
}
