% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_spacegame.R
\name{util_task_spacegame}
\alias{util_task_spacegame}
\title{util_task_spacegame: Process raw data from the Space Game (2-stage reinforcement learning task)}
\usage{
util_task_spacegame(
  sub,
  ses = 1,
  bids_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{ses}{session label used in ses-label. Default = 1}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = TRUE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function formats, cleans, and organizes spacegame data (.mat) from bids/sourcedata into rawdata for a given subject
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
Function adapted from Alaina Pearce dataBRAKEr package (https://github.com/alainapearce/dataBRAKEr/tree/main/R)
}
\examples{
\dontrun{
# process task data
bids_wd = "/Users/baf44/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/bids"
spacegame_task_pardat <- util_task_spacegame(sub = 001, ses = 1, bids_wd = bids_wd, return = TRUE)

}


}
