% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_sst.R
\name{util_task_sst}
\alias{util_task_sst}
\title{util_task_sst: Clean and organize Stop Signal Task data into BIDS rawdata}
\usage{
util_task_sst(sub_str, ses_str = "ses-1", bids_wd, overwrite = FALSE)
}
\arguments{
\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}
}
\value{
statement of task completed
}
\description{
This function formats and organizes Stop Signal Task data from bids/sourcedata into bids/rawdata for a given subject
Data from practice and behavioral runs will be organized into bids/rawdata/sub-label/beh, while data from fmri runs will be organized into bids/rawdata/sub-label/func.
}
\examples{

\dontrun{
# process task data for the Food View Task
util_task_sst(sub_str = 'sub-001', ses_str = 'ses-1', bids_wd = bids_wd)

}
}
