% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_toolbox.R
\name{util_task_toolbox}
\alias{util_task_toolbox}
\title{util_task_toolbox: Clean and organize NIH toolbox assessment data into rawdata}
\usage{
util_task_toolbox(sub, ses, bids_wd, overwrite = FALSE, return_data = TRUE)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required (integer)}

\item{ses}{session label used in ses-label (integer)}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list with 1 cleaned dataframe per run
}
\description{
This function formats and organizes NIH toolbox assessment data from bids/sourcedata into rawdata for a given subject. Assessment data includes responses to each trial in the NIH toolbox.
}
\examples{

\dontrun{
# process assessment (response) data for the NIH toolbox
sub001_toolbox_responses <- util_task_toolbox(sub = 001, ses = 1, bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}

}
