% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_untouched_to_source.R
\name{util_task_untouched_to_source}
\alias{util_task_untouched_to_source}
\title{util_task_untouched_to_source: Move task data from untouchedRaw into bids/sourcedata}
\usage{
util_task_untouched_to_source(base_wd, overwrite = FALSE)
}
\arguments{
\item{base_wd}{string with full path to base directory -- this is the directory that contains untouchedraw/ and bids/sourcedata/}

\item{overwrite}{logical indicating whether file should be overwritten in sourcedata (logical)}

\item{file}{path to file to be copied into sourcedata. Include full file path and filename (e.g., "path/to/filename.txt") (string)}

\item{sub_str}{bids-formatted subject string. e.g., "sub-001" (string)}

\item{ses_str}{bids-formatted session string. e.g., "ses-1" (string)}

\item{sourcefile_prefix}{(optional) string to prefix filename with in sourcedata (string)}
}
\description{
This function copies task data from untouchedRaw into bids/sourcedata for the following tasks: sst, foodview task, space game, pit task
}
\examples{

# organize task data in untouchedRaw into sourcedata
util_task_untouched_to_source(base_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/")

\dontrun{
}


}
