% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jsons.R
\name{write_jsons}
\alias{write_jsons}
\title{write_jsons: Write jsons created with json functions (called within proc_redcap.R)}
\usage{
write_jsons(export_dir, overwrite)
}
\arguments{
\item{export_dir}{string with absolute path to export directory (typically bids phenotype directory)}

\item{overwrite}{logical (TRUE/FALSE) to indicate if json files should be overwritten}
}
\description{
This function organizes writes json meta-data files
}
