% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_task_jsons.R
\name{write_task_jsons}
\alias{write_task_jsons}
\title{write_task_jsons: Write meta-data for tasks created with json functions (called within proc_task.R)}
\usage{
write_task_jsons(export_dir, overwrite)
}
\arguments{
\item{export_dir}{string with absolute path to export directory (typically bids rawdata directory)}

\item{overwrite}{logical (TRUE/FALSE) to indicate if json files should be overwritten}
}
\description{
This function exports json meta-data files for task data
}
