% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_foodview.R
\name{deriv_foodview}
\alias{deriv_foodview}
\title{deriv_foodview: Generate derivative databases for Foodview Task analyses}
\usage{
deriv_foodview(data)
}
\arguments{
\item{data}{a list of list of dataframes. The top-level list represents individual subjects, and each subject has a sublist of dataframes. Each dataframe contains events data for a specific run of the foodview task for a given sub. A suitable list is returned by proc_task, or can be gathered from files in bids/rawdata}

\item{return_data}{If return_data is set to TRUE, will return a list including:
\enumerate{
\item summary_long_by_cond = a long dataframe with summary data by commerical_condition (metrics calculated across runs)
\item summary_long_by_block = a long dataframe with summary data by block
}}
}
\description{
This function generates FoodView derivative databases from participant-level FoodView files
}
\examples{

\dontrun{

# process task data
base_dir = "/Users/bari/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/"
task_data <- proc_task(base_wd = base_dir, return_data = TRUE)

# get deriv foodview data from processed task data
food_view_summary <- deriv_foodview(task_data$foodview)
}
}
