% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_foodview_onsets.R
\name{deriv_foodview_onsets}
\alias{deriv_foodview_onsets}
\title{deriv_foodview_onsets: Generate onset files for foodview task}
\usage{
deriv_foodview_onsets(
  sub,
  ses = 1,
  bids_wd,
  overwrite = FALSE,
  return_data = TRUE
)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{ses}{session label used in ses-label. Default = 1}

\item{bids_wd}{string with full path to bids directory -- this directory contains rawdata/}

\item{overwrite}{logical indicating if onset files should be overwritten in /derivatives. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = TRUE}
}
\value{
If return_data is set to TRUE, will return a long dataframe with onset times for each trial_type by run
}
\description{
This function generates onset files for the foodview task formatted for analyses in AFNI. One onset file is generated for each of the following trial types: food commercial block (food_ad), toy commercial block (toy_ad), high-ED savory food image block after toy commercial (hed_savory_toy_cond), high-ED savory food image block after food commercial (hed_savory_food_cond), high-ED sweet food image block after toy commercial (hed_sweet_toy_cond), high-ED sweet food image block after food commercial (hed_sweet_food_cond), low-ED savory food image block after toy commercial (led_savory_toy_cond), low-ED savory food image block after food commercial (led_savory_food_cond), low-ED sweet food image block after toy commercial (led_sweet_toy_cond), low-ED sweet food image block after food commercial (led_sweet_food_cond)
}
\examples{

\dontrun{
# generate onset files
sub_001_fv_onsets <- deriv_foodview_onsets(sub = 001, ses = 1, bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}
}
