% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_rrv.R
\name{deriv_rrv}
\alias{deriv_rrv}
\title{deriv_rrv: Generate derivative databases for RRV analyses}
\usage{
deriv_rrv(data)
}
\arguments{
\item{data}{a list of dataframes, where 1 dataframe contains processed RRV data for 1 sub (output of util_task_rrv). A suitable list of dataframes is returned by proc_task, or can be gathered from files in bids/rawdata}

\item{return_data}{If return_data is set to TRUE, will return a list including:
\enumerate{
\item a dataframe with long summary data (by reinforcer schedule)
\item a dataframe with overall summary data
}}
}
\description{
This function generates RRV derivative databases from participant-level RRV files
}
