% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_anthro.R
\name{json_anthro}
\alias{json_anthro}
\title{json_anthro: Generates a json file for raw anthropometrics data collected during the visit}
\usage{
json_anthro()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for raw anthropometrics data collected during the visit
}
