% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_audit.R
\name{json_audit}
\alias{json_audit}
\title{json_audit: Generates a json file for the Alcohol Use Disorders Identification Test}
\usage{
json_audit()
}
\value{
A json file documenting the raw inputs and scored values for the Alcohol Use Disorders Identification Test
}
\description{
This function generates a json file for the scored Alcohol Use Disorders Identification Test and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_audit function in dataprepr and is only accurate for data collected in Study REACH
}
