% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cbq.R
\name{json_cbq}
\alias{json_cbq}
\title{json_cbq: Generates a json file for the Children\'s Behavior Questionnaire}
\usage{
json_cbq()
}
\value{
A string with data stored in JSON format containing meta-data for the Child Behavior Questionnaire
}
\description{
This function generates a json file for the scored Children\'s Behavior Questionnaire and raw participant responses.
}
