% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cchip.R
\name{json_cchip}
\alias{json_cchip}
\title{json_cchip: Generates a json file for the Community Childhood Hunger Identification Project}
\usage{
json_cchip()
}
\value{
A string with data stored in JSON format containing meta-data for the Community Childhood Hunger Identification Project
}
\description{
This function generates a json file for the scored Community Childhood Hunger Identification Project and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_cchip function in dataprepr and is only accurate for data collected in Study REACH
}
