% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cfpq.R
\name{json_cfpq}
\alias{json_cfpq}
\title{json_cfpq: Generates a json file for the Comprehensive Feeding Practices Questionnaire}
\usage{
json_cfpq()
}
\value{
A json file documenting the raw inputs and scored values for the Comprehensive Feeding Practices Questionnaire
}
\description{
This function generates a json file for the scored Comprehensive Feeding Practices Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_cfpq function in dataprepr and is only accurate for data collected in Study REACH
}
