% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cfq.R
\name{json_cfq}
\alias{json_cfq}
\title{json_cfq: Generates a json file for the Child Feeding Questionnaire}
\usage{
json_cfq()
}
\value{
A string with data stored in JSON format containing meta-data for the Child Feeding Questionnaire
}
\description{
This function generates a json file for the scored Child Feeding Questionnaire and raw participant responses.
}
