% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_debq.R
\name{json_debq}
\alias{json_debq}
\title{json_debq: Generates a json file for the Dutch Eating Behavior Questionnaire}
\usage{
json_debq()
}
\value{
A json file documenting the raw inputs and scored values for the Dutch Eating Behavior Questionnaire
}
\description{
This function generates a json file for the scored Dutch Eating Behavior Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_debq function in dataprepr and is only accurate for data collected in Study REACH
}
