% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_deriv_rrv.R
\name{json_deriv_rrv}
\alias{json_deriv_rrv}
\title{json_deriv_rrv: Generates a json file for the derivative RRV data}
\usage{
json_deriv_rrv()
}
\value{
A string with data stored in JSON format containing meta-data the summary RRV database
}
\description{
This function generates a json file the summary RRV outcomes reported in derivatives/beh_databases
}
