% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_efcr.R
\name{json_efcr}
\alias{json_efcr}
\title{json_efcr: Generates a json file for the External Food Cue Responsiveness Scale}
\usage{
json_efcr()
}
\value{
A string with data stored in JSON format containing meta-data for the External Food Cue Responsiveness Scale
}
\description{
This function generates a json file for the scored External Food Cue Responsiveness Scale and raw participant responses.
}
