% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ffbs.R
\name{json_ffbs}
\alias{json_ffbs}
\title{json_ffbs: Generates a json file for the Family Food Behavior Survey}
\usage{
json_ffbs()
}
\value{
A json file documenting the raw inputs and scored values for the Family Food Behavior Survey
}
\description{
This function generates a json file for the scored Family Food Behavior Survey and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_ffbs function in dataprepr and is only accurate for data collected in Study REACH
}
