% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_hfssm.R
\name{json_hfssm}
\alias{json_hfssm}
\title{json_hfssm: Generates a json file for the U.S. Household Food Security Survey Module}
\usage{
json_hfssm()
}
\value{
A json file documenting the raw inputs and scored values for the U.S. Household Food Security Survey Module
}
\description{
This function generates a json file for the scored U.S. Household Food Security Survey Module and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_debq function in dataprepr and is only accurate for data collected in Study REACH
}
