% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_lbc.R
\name{json_lbc}
\alias{json_lbc}
\title{json_lbc: Generates a json file for the Lifestyle Behavior Checklist}
\usage{
json_lbc()
}
\value{
A json file documenting the raw inputs and scored values for the Lifestyle Behavior Checklist
}
\description{
This function generates a json file for the scored Lifestyle Behavior Checklist and raw participant responses. This function provides accurate json files ONLY if data is processed using score_lbc function in dataprepr and is only accurate for data collected in Study BRAKE.
}
