% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_mri_vist.R
\name{json_mri_visit}
\alias{json_mri_visit}
\title{json_mri_visit: Generates a json file for MRI visit data}
\usage{
json_mri_visit()
}
\value{
A string with data stored in JSON format containing meta-data for the MRI visit data
}
\description{
This function generates a json file for MRI visit data (notes, CAMS, freddy fullness scores)
}
