% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pptq.R
\name{json_pptq}
\alias{json_pptq}
\title{json_pptq: Generates a json file for the Pictorial Personality Traits Questionnaire for Children}
\usage{
json_pptq()
}
\value{
A string with data stored in JSON format containing meta-data for the Pictorial Personality Traits Questionnaire for Children
}
\description{
This function generates a json file for the scored Pictorial Personality Traits Questionnaire for Children  and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_pptq function in dataprepr and is only accurate for data collected in Study REACH
}
