% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_puberty.R
\name{json_puberty}
\alias{json_puberty}
\title{json_puberty: Generates a json file for Pubertal Development Scale}
\usage{
json_puberty()
}
\value{
A string with data stored in JSON format containing meta-data for the Pubertal Development Score.
}
\description{
This function generates a json file for the scored Pubertal Development Scale and and raw participant responses.
This function provides accurate json files ONLY if data is child-report and adult-reported ratings are processed using score_pds function in dataprepr and combined using proc_redcap() in dataREACHr.
This is only accurate for data collected in Study BREAK
}
