% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pwlb.R
\name{json_pwlb}
\alias{json_pwlb}
\title{json_pwlb: Generates a json file for the Parent Weight Loss Behavior Questionnaire}
\usage{
json_pwlb()
}
\value{
A json file documenting the raw inputs and scored values for the Parent Weight Loss Behavior Questionnaire
}
\description{
This function generates a json file for the scored Parent Weight Loss Behavior Questionnaire and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_pwlb function in dataprepr and is only accurate for data collected in Study REACH
}
