% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pwlb.R
\name{json_pwlb}
\alias{json_pwlb}
\title{json_pwlb: Generates a json file for the Parent Weight Loss Behavior Questionnaire}
\usage{
json_pwlb()
}
\value{
A string with data stored in JSON format containing meta-data for the Parent Weight Loss Behavior Questionnaire
}
\description{
This function generates a json file for the scored Parent Weight Loss Behavior Questionnaire and raw participant responses.
}
