% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sleeplog.R
\name{json_sleeplog}
\alias{json_sleeplog}
\title{json_sleeplog: Generates a json file for the External Food Cue Responsiveness Scale}
\usage{
json_sleeplog()
}
\value{
A string with data stored in JSON format containing meta-data for the External Food Cue Responsiveness Scale
}
\description{
This function generates a json file for the scored External Food Cue Responsiveness Scale and raw participant responses.
This function provides accurate json files ONLY if data is processed using score_sleeplog function in dataprepr and is only accurate for data collected in Study REACH
}
