% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(
  visit_data_path,
  data_de_path,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{visit_data_path}{full path to the redcap visit data in bids/sourcedata/phenotype directory}

\item{data_de_path}{full path to the redcap double entry data in bids/sourcedata/phenotype directory}

\item{overwrite}{overwrite existing files (default = FALSE)}

\item{return_data}{return phenotype to console (default = FLASE)}
}
\value{
If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean raw phenotype datasets for each task
\item meta-data/.json information for each task
}
}
\description{
This function:
\enumerate{
\item Reads REDCap data (visit and double-entry) from bids/sourcedata
\item Calls util_ functions to clean and compile data in dataframes
\item Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe
\item Exports the following BIDS-compliant .tsv files into bids/phenotype:
\itemize{
\item questionnaire data - raw and scores: efcr, lbc, pss, brief2, bes, ffbs, spsrq, pwlb, tfeq, bisbas, debq, scpf, hfssm, cchip, pptq, stq, kbas, cebq, audit, cfpq, puberty, cshq, chaos, cfq
\item questionnaire data - raw only: infancy, household, fsq, hfias, sic, fhfi, pstca, pmum, rank, loc, cbq
\item compiled and researcher-entered data: demographics, intake, anthropometrics, dexa, mri_visit, sleeplog, (wasi?, notes?, updates?)
}
\item Exports bids/participants.tsv
\item Exports a .json file with meta-data for each .tsv
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
data_de_path = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids/sourcedata/phenotype/REACHDataDoubleEntry_DATA_2024-03-12_1045.csv"
visit_data_path = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids/sourcedata/phenotype/FoodMarketingResilie_DATA_2024-05-20_1520.csv"
data_de_path = "/Users/bari/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData//bids/sourcedata/phenotype/REACHDataDoubleEntry_DATA_2024-04-08_1306.csv"
visit_data_path = "/Users/bari/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData//bids/sourcedata/phenotype/FoodMarketingResilie_DATA_2024-07-17_1252.csv"
redcap_data <- proc_redcap(visit_data_path, data_de_path, return = TRUE)

}

}
