% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(base_wd, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) data should be overwritten. Default = FALSE}
}
\value{
Will return a list including:
\enumerate{
\item input_data: list with 2 dataframes with raw data (visit_data, de_data (double_entered data))
\item visit_data: list of 10 dataframes with intermediate-processed visit data (child_v1_data, child_v2_data, child_v3_data, child_v4_data, child_v5_data, parent_v1_data, parent_v2_data, parent_v3_data, parent_v4_data, parent_v5_data)
\item double_entry_data: list of dataframes with intermediate-processed double entry data
\item phenotype_data: list of dataframes with processed data that gets exported into bids/phenotype
}
}
\description{
This function:
\enumerate{
\item Reads REDCap data (visit and double-entry) from bids/sourcedata
\item Calls util_ functions to clean and compile data in dataframes
\item Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe
\item Exports the following BIDS-compliant .tsv files into bids/phenotype:
\itemize{
\item questionnaire data - raw and scores: efcr, lbc, pss, brief2, bes, ffbs, spsrq, pwlb, tfeq, bisbas, debq, scpf, hfssm, cchip, pptq, stq, kbas, cebq, audit, cfpq, puberty, cshq, chaos, cfq
\item questionnaire data - raw only: infancy, household, fsq, hfias, sic, fhfi, pstca, pmum, rank, loc, cbq
\item compiled and researcher-entered data: demographics, intake, anthropometrics, dexa, mri_visit, sleeplog, (wasi?, notes?, updates?)
}
\item Exports bids/participants.tsv
\item Exports a .json file with meta-data for each .tsv
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
redcap_data <- proc_redcap(base_wd, overwrite = FALSE, overwrite_jsons = FALSE)

}

}
