% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(visit_data_path, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{visit_data_path}{full path to the redcap visit data in bids/sourcedata/phenotype directory}

\item{overwrite}{overwrite existing files (default = FALSE)}

\item{return_data}{return phenotype to console (default = FLASE)}
}
\value{
If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean raw phenotype datasets for each task
\item meta-data/.json inforamtion for each task
}
}
\description{
This function:
\enumerate{
\item reads REDCap data from sourcedata
\item cleans data to save in BIDS format in phenotype. Produces the following files:
*
\item calls functions to create .json files for each phoneypte/x.tsv file
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{
visit_data_path = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids/sourcedata/phenotype/FoodMarketingResilie_DATA_2024-02-16_1544.csv"

phenotype_data <- proc_redcap(visit_data_path, return = TRUE)

\dontrun{
}

}
