% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(
  visit_data_path,
  data_de_path,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{visit_data_path}{full path to the redcap visit data in bids/sourcedata/phenotype directory}

\item{data_de_path}{full path to the redcap double entry data in bids/sourcedata/phenotype directory}

\item{overwrite}{overwrite existing files (default = FALSE)}

\item{return_data}{return raw and processed data to console (default = FALSE)}
}
\value{
If return_data is set to TRUE, will return a list including:
\enumerate{
\item input_data: list with 2 dataframes with raw data (visit_data, de_data (double_entered data))
\item visit_data: list of 10 dataframes with intermediate-processed visit data (child_v1_data, child_v2_data, child_v3_data, child_v4_data, child_v5_data, parent_v1_data, parent_v2_data, parent_v3_data, parent_v4_data, parent_v5_data)
\item double_entry_data: list of dataframes with intermediate-processed double entry data
\item phenotype_data: list of dataframes with processed data that gets exported into bids/phenotype
}
}
\description{
This function:
\enumerate{
\item Reads REDCap data (visit and double-entry) from bids/sourcedata
\item Calls util_ functions to clean and compile data in dataframes
\item Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe
\item Exports the following BIDS-compliant .tsv files into bids/phenotype:
\itemize{
\item questionnaire data - raw and scores: efcr, lbc, pss, brief2, bes, ffbs, spsrq, pwlb, tfeq, bisbas, debq, scpf, hfssm, cchip, pptq, stq, kbas, cebq, audit, cfpq, puberty, cshq, chaos, cfq
\item questionnaire data - raw only: infancy, household, fsq, hfias, sic, fhfi, pstca, pmum, rank, loc, cbq
\item compiled and researcher-entered data: demographics, intake, anthropometrics, dexa, mri_visit, sleeplog, (wasi?, notes?, updates?)
}
\item Exports bids/participants.tsv
\item Exports a .json file with meta-data for each .tsv
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
data_de_path = "/Users/baf44/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/bids/sourcedata/phenotype/REACHDataDoubleEntry_DATA_2024-08-14_1521.csv"
visit_data_path = "/Users/baf44/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/bids/sourcedata/phenotype/FoodMarketingResilie_DATA_2024-08-14_1522.csv"
redcap_data <- proc_redcap(visit_data_path, data_de_path, return = TRUE)

}

}
