% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_task.R
\name{proc_task}
\alias{proc_task}
\title{proc_task: Process task data in untouchedRaw into bids}
\usage{
proc_task(base_wd, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{base_wd}{full path to directory that contains untouchedRaw/ and bids/}

\item{overwrite}{overwrite existing files (default = FALSE)}

\item{return_data}{return phenotype to console (default = FLASE)}
}
\value{
If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean datasets for each task
\item meta-data/.json inforamtion for each task
}
}
\description{
This function:
\enumerate{
\item copies task data from untouchedRaw into sourcedata
\item cleans sourcedata to save in BIDS format in rawdata. Produces the following files:
*
\item calls functions to create .json files for task data
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{
\dontrun{
task_data <- proc_task(base_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/", return_data = TRUE)

}

}
