% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_task.R
\name{proc_task}
\alias{proc_task}
\title{proc_task: Process task data from untouchedRaw to bids/rawdata}
\usage{
proc_task(
  base_wd,
  overwrite_sourcedata = FALSE,
  overwrite_rawdata = FALSE,
  overwrite_rawdata_vector,
  overwrite_jsons = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains untouchedRaw/ and bids/}

\item{overwrite_sourcedata}{(logical) overwrite files in sourcedata for all tasks (default = FALSE)}

\item{overwrite_rawdata}{(logical) overwrite files in rawdata for all tasks (default = FALSE)}

\item{overwrite_rawdata_vector}{(vector) vector with names of tasks for which rawdata should be overwritten. Options include: c("sst", "foodview", "nih_toolbox", "rrv", "pit").}

\item{overwrite_jsons}{overwrite existing jsons in rawdata. Applies to all tasks (default = FALSE) (logical)}

\item{return_data}{return BIDS data (i.e., data ready for bids/rawdata) for each task to console (default = FLASE) (logical)}
}
\value{
If return_data is set to TRUE, will return a list including:
\enumerate{
\item lists with cleaned dataframes for each subject for each task
\item meta-data information for each task stored in JSON format
}
}
\description{
This function:
\enumerate{
\item copies task data from untouchedRaw into bids/sourcedata for all tasks (food view, pit, sst, nih toolbox, spacegame), using util_task_untouched_to_source(all_tasks = TRUE)
\item processes task sourcedata and exports cleaned dataframes into bids/rawdata for the following tasks: rrv, sst, foodview nih-toolbox, pit, using task-specific util_task_{task-name} functions
\item exports JSON meta-data files for tasks organized into rawdata (rrv, sst, foodview), using write_task_jsons()
}
}
\details{
To use this function, the correct path to the directory containing untouchedRaw/ and bids/ must be supplied to base_wd
}
\examples{
\dontrun{

# set base_wd
base_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/"

# process task data without overwriting any existing files, return processed data
task_data <- proc_task(base_wd = base_wd, return_data = TRUE)

# process task data and overwrite rawdata for foodview task only
proc_task(base_wd = base_wd, overwrite_rawdata_vector = c("foodview"))

# process task data and overwrite all sourcedata, rawdata, and meta-data
proc_task(base_wd = base_wd, overwrite_sourcedata = TRUE, overwrite_rawdata = TRUE, overwrite_jsons = TRUE)

}

}
