% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_task_derivs.R
\name{proc_task_derivs}
\alias{proc_task_derivs}
\title{proc_task_derivs: Exports derivative databases with task beh summary metrics and associated meta-data}
\usage{
proc_task_derivs(task_data, export_dir)
}
\arguments{
\item{task_data}{output returned by proc_task()}

\item{export_dir}{string with absolute path to export directory (typically behavioral summary directory)}
}
\description{
This function calls deriv_{task} functions and exports TSVs with summary behavioral metrics and associated meta-data
}
\examples{

\dontrun{

# process task data
base_dir = "/Users/bari/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/"
task_data <- proc_task(base_wd = base_dir, return_data = TRUE)

# export derivative databases
proc_task_derivs(task_data, "path/to/export/")
}

}
