% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_task_derivs.R
\name{proc_task_derivs}
\alias{proc_task_derivs}
\title{proc_task_derivs: Generate derivative task data from individual files in rawdata}
\usage{
proc_task_derivs(
  base_wd,
  overwrite = FALSE,
  proc_source = proc_source,
  task_list = "all",
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{proc_source}{whether to processes the raw data using proc_tasks.R. Default = FALSE}

\item{task_list}{tasks to process. Options include 'all' to process all task data or a list of the following:\itemize{
\item{'foodview' - fMRI Food Viewing task}
\item{'nihtoolbox' - NIH Toolbox data}
\item{'pit' - Pavlovian Instrumental Transfer task data}
\item{'rrv' - Relative Reinforcing Value of Food task}
\item{'spacegame' - Space Game data (need to finish processing in Matlab)}
\item{'sst' - fMRI Stop-Signal Task data}
}}
}
\description{
This function: \itemize{
\item{1) processes task data from rawdata to generate derivative data in bids/derivatives}
\item{2) generate json files for derivative task files}
\item{3) write data and json files}
}
}
\examples{

\dontrun{
# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
proc_task(base_wd = base_wd, task_list = c('spacegame', 'nih_toolbox'))

}


}
