% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_tasks.R
\name{proc_tasks}
\alias{proc_tasks}
\title{proc_tasks: Process task data from untouchedRaw to create bids compliant files}
\usage{
proc_tasks(base_wd, overwrite = FALSE, task_list)
}
\arguments{
\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}

\item{task_list}{tasks to process. Options include 'all' to process all task data or a list of the following:\itemize{
\item{'foodview' - fMRI Food Viewing task}
\item{'nihtoolbox' - NIH Toolbox data}
\item{'pit' - Pavlovian Instrumental Transfer task data}
\item{'rrv' - Relative Reinforcing Value of Food task}
\item{'spacegame' - Space Game data (need to finish processing in Matlab)}
\item{'sst' - fMRI Stop-Signal Task data}
}}
}
\description{
This function: \itemize{
\item{1) copies task data from untouchedRaw into bids/sourcedata for all tasks (food view, pit, sst, nih toolbox, spacegame), using util_task_untouched_to_source(all_tasks = TRUE)}
\item{2) processes task sourcedata and exports cleaned dataframes into bids/rawdata for the following tasks: rrv, sst, foodview nih-toolbox, pit, spacegame, using task-specific util_task_{task-name} functions}
\item{3) exports JSON meta-data files for tasks organized into rawdata (rrv, sst, foodview), using write_task_jsons()}
}
}
\examples{

\dontrun{
# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
proc_task(base_wd = base_wd, task_list = c('spacegame', 'nih_toolbox'))

}


}
