% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrv_parse_text.R
\name{rrv_parse_text}
\alias{rrv_parse_text}
\title{rrv_parse_text: Generate tabular data from rrv text files}
\usage{
rrv_parse_text(rrv_file)
}
\arguments{
\item{rrv_file}{string with full path to rrv file}
}
\description{
This function generates and returns dataframes (game, summary) from rrv text file
}
\examples{

\dontrun{

# process task data for the Food View Task
file_name = "/Users/baf44/projects/Keller_Marketing/ParticipantData/untouchedRaw/rrv_task/REACH_060/060.txt"
rrv_csv <- rrv_parse_text(sub = 60, rrv_file = file_name)

}


}
