% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrv_parse_text.R
\name{rrv_parse_text}
\alias{rrv_parse_text}
\title{rrv_parse_text: Generate tabular data from rrv text files}
\usage{
rrv_parse_text(rrv_file, participant_id)
}
\arguments{
\item{rrv_file}{(character string) full path to rrv file}

\item{participant_id}{(character string) bids-compliant subject ID}
}
\description{
This function generates and returns 1 dataframe with data from input RRV file
}
\examples{

\dontrun{

# process task data for the Food View Task
file_name = "/Users/baf44/projects/Keller_Marketing/ParticipantData/untouchedRaw/rrv_task/REACH_060/rrv_060.txt"
rrv_csv <- rrv_parse_text(rrv_file = file_name, participant_id = "sub-060")

}


}
