% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_calc_intake.R
\name{util_calc_intake}
\alias{util_calc_intake}
\title{util_calc_intake: Computes intake variables (called within util_redcap_de.R)}
\usage{
util_calc_intake(stacked_intake)
}
\arguments{
\item{stacked_intake}{Intake data (i.e., pre and post weights) extracted double entry data}
}
\value{
A dataframe of stacked_intake (input) variables and calculated intake variables
}
\description{
This function computes intake variables
}
