% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_kbas_data.R
\name{util_format_kbas_data}
\alias{util_format_kbas_data}
\title{util_format_kbas_data: format Kids Brand Awareness Survey data data}
\usage{
util_format_kbas_data(kbas_data)
}
\arguments{
\item{kbas_data}{Kids Brand Awareness Survey extracted from data from REDCap events}
}
\value{
formatted kbas_data
}
\description{
This function formats Kids Brand Awareness Survey data
}
\examples{

# process kbas data
kbas_data_formatted <- util_format_kbas_data(kbas_data)

}
\seealso{
\code{\link[=util_redcap_child_v1]{util_redcap_child_v1()}}, \code{\link[=util_redcap_child_v5]{util_redcap_child_v5()}}
}
