% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_rank_data.R
\name{util_format_rank_data}
\alias{util_format_rank_data}
\title{util_format_rank_data: process Ranking Food Item Questionnaire data}
\usage{
util_format_rank_data(rank_data)
}
\arguments{
\item{rank_data}{rank extracted from data from REDCap events util_redcap_parent_v1}
}
\description{
This function process Ranking Food Item Questionnaire data
}
\seealso{
\code{\link[=util_redcap_parent_v1]{util_redcap_parent_v1()}}
}
