% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_anthro.R
\name{util_merged_anthro}
\alias{util_merged_anthro}
\title{util_merged_anthro: merges and formats anthropometrics data}
\usage{
util_merged_anthro(visit1_anthro, visit5_anthro, merged_household, date_data)
}
\arguments{
\item{visit1_anthro}{visit 1 anthropometrics data.frame}

\item{visit5_anthro}{visit 2 anthropometrics data.frame}

\item{merged_household}{merged household data.frame}

\item{date_data}{data.frame with all visit date information}
}
\description{
This function merges anthropometrics data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_anthro_data <- util_merged_anthro(child_v1_data$anthro_data$data, child_v5_data$anthro_data$data, merged_qs_list$household_all, date_data)

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}, \code{\link[=util_merge_questionnaires]{util_merge_questionnaires()}}
}
