% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v1.R
\name{util_redcap_child_v1}
\alias{util_redcap_child_v1}
\title{util_redcap_child_v1: Organize child visit 1 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_child_v1(data)
}
\arguments{
\item{data}{data from REDCap event child_visit_1_arm_1}
}
\value{
Will return a list including:
\itemize{
\item{clean raw child visit 1 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{visit_data_child}
\item{food_paradigm_info}
\item{freddy_data}
\item{intake_data}
\item{liking_data}
\item{kbas_data}
\item{stq_data}
\item{anthro_data}
}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_1_arm_1
}
\examples{

# process REDCap data
child_visit1_list <- util_redcap_child_v1(data)

\dontrun{
}

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
