% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v2.R
\name{util_redcap_child_v2}
\alias{util_redcap_child_v2}
\title{util_redcap_child_v2: Organize child visit 2 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_child_v2(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event child_visit_2_arm_1}

\item{return_data}{If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean raw child visit 1 datasets
\item meta-data/.json for each dataset
}}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_2_arm_1
}
