% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child_v4.R
\name{util_redcap_child_v4}
\alias{util_redcap_child_v4}
\title{util_redcap_child_v4: Organize child visit 4 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_child_v4(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event child_visit_4_arm_1}

\item{return_data}{If return_data is set to TRUE, will return a list of dataframes including: visit_data_child, food_paradigm_info, eah_wanting, freddy_data, intake_data, loc_data, pptq_data, sic_data}
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_4_arm_1
}
