% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_de.R
\name{util_redcap_de}
\alias{util_redcap_de}
\title{util_redcap_de: Organize double-entry data from REDCap}
\usage{
util_redcap_de(redcap_api = FALSE, redcap_de_data, date_data)
}
\arguments{
\item{redcap_api}{(logical) execute REDCap API. Default = FALSE.}

\item{redcap_de_data}{REDCap double-entry data from a prior API call}

\item{date_data}{date data for REDCap visit

Will return a list including:
\itemize{
\item{clean raw and scored parent visit 1 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{demo_data}
\item{infancy_data}
\item{household_data}
\item{rank_data}
\item{puberty_data}
\item{cfq_data}
\item{cebq_data}
\item{efcr_data}
\item{chaos_data}
\item{pss_data}
\item{lbc_data}
}}
}
\value{
Will return a list including data that has been double-entered and checked along with the metadata for:
\itemize{
\item{dxa_v1}
\item{dxa_v5}
\item{intake_v1}
\item{intake_v3}
\item{intake_v4}
\item{intake_v5}
}
}
\description{
This function organizes REDCap double entry data data
}
