% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent_v1.R
\name{util_redcap_parent_v1}
\alias{util_redcap_parent_v1}
\title{util_redcap_parent_v1: Organize parent visit 1 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_parent_v1(data, v1_date_data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event parent_visit_1_arm_1}

\item{v1_date_data}{dataset that includes ID variable and date of first visit}

\item{return_data}{If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean raw parent 1 datasets
\item meta-data/.json for each dataset
}}
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_1_arm_1
}
