% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent_v2.R
\name{util_redcap_parent_v2}
\alias{util_redcap_parent_v2}
\title{util_redcap_parent_v2: Organize parent visit 2 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_parent_v2(data, agesex_data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event parent_visit_2_arm_1e'}

\item{agesex_data}{dataframe with columns: 'participant_id', 'v2_age', 'sex' -- can have additional columns as long as these are in there; rerequired for scoring the BRIEF2}

\item{return_data}{If return_data is set to TRUE, will return a list including:
\enumerate{
\item clean raw parent 1 datasets
\item meta-data/.json for each dataset
}}
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_2_arm_1
}
