% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_foodview.R
\name{util_task_foodview}
\alias{util_task_foodview}
\title{util_task_foodview: Clean and organize Food View task data into BIDS rawdata}
\usage{
util_task_foodview(
  sub,
  ses = 1,
  bids_wd,
  overwrite = FALSE,
  return_data = TRUE
)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{ses}{session label used in ses-label. Default = 1}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list with 1 cleaned dataframe per run
}
\description{
This function formats and organizes Food View task data from bids/sourcedata into bids/rawdata for a given subject
}
\examples{

\dontrun{
# process task data for the Food View Task
list_of_cleaned_data <- util_task_foodview_orgraw(sub = 001, ses = 1, bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}


}
