% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_foodview.R
\name{util_task_foodview}
\alias{util_task_foodview}
\title{util_task_foodview: Clean and organize Food View task data into BIDS rawdata}
\usage{
util_task_foodview(sub, ses, bids_wd, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{sub}{subject label used in sub-\if{html}{\out{<label>}}. Leading zeros not required}

\item{ses}{session label used in ses-\if{html}{\out{<label>}}}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset for each run
}
\description{
This function processes foodview task sourcedata into BIDS and exports it into rawdata
}
\examples{

# process task data for the Food View Task
test <- util_task_foodview_orgraw(sub = 001, ses = 1, data_path, return = TRUE)

\dontrun{
}


}
