% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_pit.R
\name{util_task_pit}
\alias{util_task_pit}
\title{util_task_pit: Clean and organize PIT data into BIDS rawdata}
\usage{
util_task_pit(sub_str, ses_str, base_wd, overwrite = FALSE)
}
\arguments{
\item{sub_str}{(string) bids-formatted session string. e.g., 'sub-001'}

\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{base_wd}{(string) full path to directory that contains both the untouchedRaw and bids directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten}
}
\value{
statement of task completed
}
\description{
This function formats and organizes RRV data from bids/sourcedata into bids/rawdata for a given subject
}
\examples{

\dontrun{
# process task data for the Food View Task
pit_data <- util_task_pit(sub_str = 'sub-001', ses_str = 'ses-1', base_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}
}
