% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_rrv.R
\name{util_task_rrv}
\alias{util_task_rrv}
\title{util_task_rrv: Clean and organize RRV data into BIDS rawdata}
\usage{
util_task_rrv(sub, ses = 1, bids_wd, overwrite = FALSE, return_data = TRUE)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{ses}{session label used in ses-label. Default = 1}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = TRUE}
}
\value{
If return_data is set to TRUE, will return a dataframe with cleaned data
}
\description{
This function formats and organizes RRV data from bids/sourcedata into bids/rawdata for a given subject
If subject has RRV text file, rrv_parse_text() will be called. If subject does not have text file but has game.csv, game.csv will processed to generate a dataframe like those generated by rrv_parse_text()
}
\examples{

\dontrun{
# process task data for the Food View Task
rrv_data <- util_task_rrv(sub = 001, ses = 1, bids_wd = "/Users/baf44/projects/Keller_Marketing/ParticipantData/bids", return = TRUE)

}
}
