% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_spacegame.R
\name{util_task_spacegame}
\alias{util_task_spacegame}
\title{util_task_spacegame: Process raw data from the Space Game (2-stage reinforcement learning task)}
\usage{
util_task_spacegame(sub_str, ses_str = "ses-1", bids_wd, overwrite = FALSE)
}
\arguments{
\item{ses_str}{(string) bids-formatted session string. e.g., 'ses-1'}

\item{bids_wd}{string with full path to bids directory that contains the sourcedata/ and rawdata/ directories}

\item{overwrite}{(logical) logical indicating whether file should be overwritten in sourcedata}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function formats, cleans, and organizes spacegame data (.mat) from bids/sourcedata into rawdata for a given subject
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
Function adapted from Alaina Pearce dataBRAKEr package (https://github.com/alainapearce/dataBRAKEr/tree/main/R)
}
\examples{
\dontrun{
# process task data

spacegame_task_pardat <- util_task_spacegame(sub_str = 'sub-001', ses_str = 'ses-1', bids_wd = bids_wd)

}


}
