% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_sst.R
\name{util_task_sst}
\alias{util_task_sst}
\title{util_task_sst: Clean and organize Stop Signal Task data into BIDS rawdata}
\usage{
util_task_sst(sub, ses = 1, bids_wd, overwrite = FALSE, return_data = TRUE)
}
\arguments{
\item{sub}{subject label used in sub-label. Leading zeros not required}

\item{ses}{session label used in ses-label. Default = 1}

\item{bids_wd}{string with full path to bids directory -- this is the directory that contains sourcedata/ and rawdata/}

\item{overwrite}{logical indicating if data should be overwritten in /rawdata. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = TRUE}
}
\value{
If return_data is set to TRUE, will return a list with 1 cleaned dataframe per run
}
\description{
This function formats and organizes Stop Signal Task data from bids/sourcedata into bids/rawdata for a given subject
Data from practice and behavioral runs will be organized into bids/rawdata/sub-label/beh, while data from fmri runs will be organized into bids/rawdata/sub-label/func.
}
\examples{

\dontrun{
bids_wd = "/Users/bari/Library/CloudStorage/OneDrive-ThePennsylvaniaStateUniversity/b-childfoodlab_Shared/Active_Studies/MarketingResilienceRO1_8242020/ParticipantData/bids"
# process task data for the Food View Task
list_of_cleaned_data <- util_task_sst(sub = 001, ses = 1, bids_wd = bids_wd, return = TRUE)

}
}
