% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phenotype_jsons.R
\name{write_phenotype_jsons}
\alias{write_phenotype_jsons}
\title{write_phenotype_jsons: Write meta-data for phenotype data created with json functions}
\usage{
write_phenotype_jsons(export_dir, overwrite)
}
\arguments{
\item{export_dir}{string with absolute path to export directory (typically bids phenotype directory)}

\item{overwrite}{logical (TRUE/FALSE) to indicate if json files should be overwritten}
}
\description{
This function exports json meta-data files for phenotype data
This function is not currently called as part of proc_redcap() as write_phenotype_data() is called instead to export both .tsv and associated meta-data
}
